opendal

Module blocking

Available on crate feature blocking only.

blocking module provides blocking APIs for OpenDAL.

Structs

BufferIterator

BufferIterator is an iterator of buffers.

Deleter

BlockingDeleter is designed to continuously remove content from storage.

Lister

BlockingLister is designed to list entries at given path in a blocking manner.

Operator

Use OpenDAL in blocking context.

Reader

BlockingReader is designed to read data from given path in an blocking manner.

StdBytesIterator

StdIterator is the adapter of Iterator for [BlockingReader][crate::BlockingReader].

StdReader

StdReader is the adapter of Read, Seek and BufRead for [BlockingReader][crate::BlockingReader].

StdWriter

StdWriter is the adapter of std::io::Write for [BlockingWriter].

Writer

BlockingWriter is designed to write data into given path in an blocking manner.