opendal::docs

Module performance

Available on docsrs only.

OpenDAL Performance Guide

OpenDAL is designed in a zero-cost abstraction way, which means that users won’t pay for the abstraction cost if they don’t use it. But this can also means that users can’t maximize the performance of OpenDAL if they don’t know how to use it.

This document will introduce some tips to improve the performance of OpenDAL.

Modules

concurrent_write

Concurrent Write

http_optimization

HTTP Optimization