Module opendal::docs::rfcs

RFCs power OpenDAL’s development.

To add new features and big refactors:

Start a new RFCs with the template rfc_0000_example.

Submit PR and assign the RFC number with the PR number.

Request reviews from OpenDAL maintainers.

Create a tracking issue and update links in RFC after approval.

Some useful tips:

Start a pre-propose in discussion to communicate quickly.

The proposer of RFC may not be the same person as the implementor. Try to include enough information in RFC itself.

Modules

rfc_0000_example

Proposal Name: (fill me in with a unique ident, my_awesome_feature)Start Date: (fill me in with today’s date, YYYY-MM-DD)RFC PR: apache/incubator-opendal#0000Tracking Issue: apache/incubator-opendal#0000

rfc_0041_object_native_api

Proposal Name: object_native_apiStart Date: 2022-02-18RFC PR: apache/incubator-opendal#41Tracking Issue: apache/incubator-opendal#35

rfc_0044_error_handle

Proposal Name: error_handleStart Date: 2022-02-23RFC PR: apache/incubator-opendal#44Tracking Issue: apache/incubator-opendal#43

rfc_0057_auto_region

Proposal Name: auto_regionStart Date: 2022-02-24RFC PR: apache/incubator-opendal#57Tracking Issue: apache/incubator-opendal#58

rfc_0069_object_stream

Proposal Name: object_streamStart Date: 2022-02-25RFC PR: apache/incubator-opendal#69Tracking Issue: apache/incubator-opendal#69

rfc_0090_limited_reader

Proposal Name: limited_readerStart Date: 2022-03-02RFC PR: apache/incubator-opendal#0090Tracking Issue: apache/incubator-opendal#0090

rfc_0112_path_normalization

Proposal Name: path-normalizationStart Date: 2022-03-08RFC PR: apache/incubator-opendal#112Tracking Issue: apache/incubator-opendal#112

rfc_0191_async_streaming_io

Proposal Name: async_streaming_ioStart Date: 2022-03-28RFC PR: apache/incubator-opendal#191Tracking Issue: apache/incubator-opendal#190

rfc_0203_remove_credential

Proposal Name: remove_credentialStart Date: 2022-04-02RFC PR: apache/incubator-opendal#203Tracking Issue: apache/incubator-opendal#203

rfc_0221_create_dir

Proposal Name: create-dirStart Date: 2022-04-06RFC PR: apache/incubator-opendal#221Tracking Issue: apache/incubator-opendal#222

rfc_0247_retryable_error

Proposal Name: retryable_errorStart Date: 2022-04-12RFC PR: apache/incubator-opendal#247Tracking Issue: apache/incubator-opendal#248

rfc_0293_object_id

Proposal Name: object_idStart Date: 2022-05-27RFC PR: apache/incubator-opendal#293Tracking Issue: apache/incubator-opendal#294

rfc_0337_dir_entry

Proposal Name: dir_entryStart Date: 2022-06-08RFC PR: apache/incubator-opendal#337Tracking Issue: apache/incubator-opendal#338

rfc_0409_accessor_capabilities

Proposal Name: accessor_capabilitiesStart Date: 2022-06-29RFC PR: apache/incubator-opendal#409Tracking Issue: apache/incubator-opendal#410

rfc_0413_presign

Proposal Name: presignStart Date: 2022-06-30RFC PR: apache/incubator-opendal#0413Tracking Issue: apache/incubator-opendal#394

rfc_0423_command_line_interface

Proposal Name: command_line_interfaceStart Date: 2022-07-08RFC PR: apache/incubator-opendal#423Tracking Issue: apache/incubator-opendal#422

rfc_0429_init_from_iter

Proposal Name: init_from_iterStart Date: 2022-07-10RFC PR: apache/incubator-opendal#429Tracking Issue: apache/incubator-opendal#430

rfc_0438_multipart

Proposal Name: multipartStart Date: 2022-07-11RFC PR: apache/incubator-opendal#438Tracking Issue: apache/incubator-opendal#439

rfc_0443_gateway

Proposal Name: gatewayStart Date: 2022-07-18RFC PR: apache/incubator-opendal#443Tracking Issue: apache/incubator-opendal#444

rfc_0501_new_builder

Proposal Name: new_builderStart Date: 2022-08-03RFC PR: apache/incubator-opendal#501Tracking Issue: apache/incubator-opendal#502

rfc_0554_write_refactor

Proposal Name: write_refactorStart Date: 2022-08-22RFC PR: apache/incubator-opendal#554Tracking Issue: apache/incubator-opendal#555

rfc_0561_list_metadata_reuse

Proposal Name: list_metadata_reuseStart Date: 2022-08-23RFC PR: apache/incubator-opendal#561Tracking Issue: apache/incubator-opendal#570

rfc_0599_blocking_api

Proposal Name: blocking_apiStart Date: 2022-08-30RFC PR: apache/incubator-opendal#599Tracking Issue: apache/incubator-opendal#601

rfc_0623_redis_service

Proposal Name: redis_serviceStart Date: 2022-08-31RFC PR: apache/incubator-opendal#0623Tracking Issue: apache/incubator-opendal#641

rfc_0627_split_capabilities

Proposal Name: split-capabilitiesStart Date: 2022-09-04RFC PR: apache/incubator-opendal#627Tracking Issue: apache/incubator-opendal#628

rfc_0661_path_in_accessor

Proposal Name: path_in_accessorStart Date: 2022-09-12RFC PR: apache/incubator-opendal#661Tracking Issue: apache/incubator-opendal#662

rfc_0793_generic_kv_services

Proposal Name: generic-kv-servicesStart Date: 2022-10-03RFC PR: apache/incubator-opendal#793Tracking Issue: apache/incubator-opendal#794

rfc_0926_object_reader

Proposal Name: object_readerStart Date: 2022-11-13RFC PR: apache/incubator-opendal#926Tracking Issue: apache/incubator-opendal#927

rfc_0977_refactor_error

Proposal Name: refactor-errorStart Date: 2022-11-21RFC PR: apache/incubator-opendal#977Tracking Issue: apache/incubator-opendal#976

rfc_1085_object_handler

Proposal Name: object_handlerStart Date: 2022-12-19RFC PR: apache/incubator-opendal#1085Tracking Issue: apache/incubator-opendal#1085

rfc_1391_object_metadataer

Proposal Name: object_metadataerStart Date: 2023-02-21RFC PR: apache/incubator-opendal#1391Tracking Issue: apache/incubator-opendal#1393

rfc_1398_query_based_metadata

Proposal Name: query_based_metadataStart Date: 2022-02-22RFC PR: apache/incubator-opendal#1398Tracking Issue: apache/incubator-opendal#1398

rfc_1420_object_writer

Proposal Name: object_writerStart Date: 2023-02-27RFC PR: apache/incubator-opendal#1420Tracking Issue: apache/incubator-opendal#1421

rfc_1477_remove_object_concept

Proposal Name: remove_object_conceptStart Date: 2023-03-05RFC PR: apache/incubator-opendal#1477Tracking Issue: apache/incubator-opendal#0000

rfc_1735_operation_extension

Proposal Name: operation_extensionStart Date: 2023-03-23RFC PR: apache/incubator-opendal#1735Tracking Issue: apache/incubator-opendal#1738

rfc_2083_writer_sink_api

Proposal Name: writer_sink_apiStart Date: 2023-04-23RFC PR: apache/incubator-opendal#2083Tracking Issue: apache/incubator-opendal#2084

rfc_2133_append_api

Proposal Name: append_apiStart Date: 2023-04-26RFC PR: apache/incubator-opendal#2133Tracking Issue: apache/incubator-opendal#2163

rfc_2299_chain_based_operator_api

Proposal Name: chain_based_operator_apiStart Date: 2023-05-23RFC PR: apache/incubator-opendal#2299Tracking Issue: apache/incubator-opendal#2300