Module opendal::executors

executors module provides implementations for the Execute trait for widely used runtimes.

Every executor will be hide behind the feature like executors-xxx. Users can switch or enable the executors they want by enabling the corresponding feature. Also, users can provide their own executor by implementing the Execute trait directly.

Structs

TokioExecutorexecutors-tokio

Executor that uses the [tokio::task::spawn] to execute futures.