opendal::layers

Trait LoggingInterceptor

pub trait LoggingInterceptor:
    Debug
    + Clone
    + Send
    + Sync
    + Unpin
    + 'static {
    // Required method
    fn log(
        &self,
        info: &AccessorInfo,
        operation: Operation,
        context: &[(&str, &str)],
        message: &str,
        err: Option<&Error>,
    );
}

LoggingInterceptor is used to intercept the log.

Required Methods

fn log( &self, info: &AccessorInfo, operation: Operation, context: &[(&str, &str)], message: &str, err: Option<&Error>, )

Everytime there is a log, this function will be called.

Inputs

info: The service’s access info.

operation: The operation to log.

context: Additional context of the log like path, etc.

message: The log message.

err: The error to log.

Note

Users should avoid calling resource-intensive operations such as I/O or network functions here, especially anything that takes longer than 10ms. Otherwise, Opendal could perform unexpectedly slow.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.