Type Alias opendal::operator_futures::FutureList

pub type FutureList<F> = OperatorFuture<OpList, Vec<Entry>, F>;

Future that generated by Operator::list_with or Operator::lister_with.

Users can add more options by public functions provided by this struct.

Aliased Type

struct FutureList<F> { /* private fields */ }

Implementations

impl<F: Future<Output = Result<Vec<Entry>>>> FutureList<F>

pub fn limit(self, v: usize) -> Self

The limit passed to underlying service to specify the max results that could return per-request.

Users could use this to control the memory usage of list operation.

pub fn start_after(self, v: &str) -> Self

The start_after passes to underlying service to specify the specified key to start listing from.

pub fn recursive(self, v: bool) -> Self

The recursive is used to control whether the list operation is recursive.

If false, list operation will only list the entries under the given path.

If true, list operation will list all entries that starts with given path.

Default to false.

pub fn metakey(self, v: impl Into<FlagSet<Metakey>>) -> Self

Metakey is used to control which meta should be returned.

Lister will make sure the result for specified meta is known:

Some(v) means exist.

None means services doesn’t have this meta.

The default metakey is Metakey::Mode.

pub fn concurrent(self, v: usize) -> Self

Concurrent is used to control the number of concurrent stat requests.

If concurrent is set to <=1, the lister will perform stat requests sequentially.

The default concurrent is 1.

pub fn version(self, v: bool) -> Self

The version is used to control whether the object versions should be returned.

If false, list operation will not return with object versions

If true, list operation will return with object versions if object versioning is supported by the underlying service

Default to false