Type Alias opendal::operator_futures::FuturePresignStat

pub type FuturePresignStat<F> = OperatorFuture<(OpStat, Duration), PresignedRequest, F>;

Future that generated by Operator::presign_stat_with.

Users can add more options by public functions provided by this struct.

Aliased Type

struct FuturePresignStat<F> { /* private fields */ }

Implementations

impl<F: Future<Output = Result<PresignedRequest>>> FuturePresignStat<F>

pub fn override_content_disposition(self, v: &str) -> Self

Sets the content-disposition header that should be sent back by the remote read operation.

pub fn override_cache_control(self, v: &str) -> Self

Sets the cache-control header that should be sent back by the remote read operation.

pub fn override_content_type(self, v: &str) -> Self

Sets the content-type header that should be sent back by the remote read operation.

pub fn if_match(self, v: &str) -> Self

Set the If-Match of the option

pub fn if_none_match(self, v: &str) -> Self

Set the If-None-Match of the option