Module opendal::raw::adapters::kv

Providing Key Value Adapter for OpenDAL.

Any services that implement Adapter can be used an OpenDAL Service.

Structs

Backend

Backend of kv service. If the storage service is one k-v-like service, it should implement this kv Backend by right.

Info

Info for this key value accessor.

Traits

Adapter

KvAdapter is the adapter to underlying kv services.

Scan

Scan is the async iterator returned by Adapter::scan.

Type Aliases

Scanner

A type-erased wrapper of Scan