Trait opendal::raw::adapters::kv::Adapter

pub trait Adapter: Send + Sync + Debug + Unpin + 'static {
    // Required methods
    fn metadata(&self) -> Metadata;
    fn get<'life0, 'life1, 'async_trait>(
        &'life0 self,
        path: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Option<Buffer>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn set<'life0, 'life1, 'async_trait>(
        &'life0 self,
        path: &'life1 str,
        value: Buffer
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn delete<'life0, 'life1, 'async_trait>(
        &'life0 self,
        path: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;

    // Provided methods
    fn blocking_get(&self, path: &str) -> Result<Option<Buffer>> { ... }
    fn blocking_set(&self, path: &str, value: Buffer) -> Result<()> { ... }
    fn blocking_delete(&self, path: &str) -> Result<()> { ... }
    fn scan<'life0, 'life1, 'async_trait>(
        &'life0 self,
        path: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<Vec<String>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
    fn blocking_scan(&self, path: &str) -> Result<Vec<String>> { ... }
    fn append<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        path: &'life1 str,
        value: &'life2 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait { ... }
    fn blocking_append(&self, path: &str, value: &[u8]) -> Result<()> { ... }
}

KvAdapter is the adapter to underlying kv services.

By implement this trait, any kv service can work as an OpenDAL Service.

Required Methods

fn metadata(&self) -> Metadata

Return the metadata of this key value accessor.

fn get<'life0, 'life1, 'async_trait>( &'life0 self, path: &'life1 str ) -> Pin<Box<dyn Future<Output = Result<Option<Buffer>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get a key from service.

return Ok(None) if this key is not exist.

fn set<'life0, 'life1, 'async_trait>( &'life0 self, path: &'life1 str, value: Buffer ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Set a key into service.

fn delete<'life0, 'life1, 'async_trait>( &'life0 self, path: &'life1 str ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Delete a key from service.

return Ok(()) even if this key is not exist.

Provided Methods

fn blocking_get(&self, path: &str) -> Result<Option<Buffer>>

The blocking version of get.

fn blocking_set(&self, path: &str, value: Buffer) -> Result<()>

The blocking version of set.

fn blocking_delete(&self, path: &str) -> Result<()>

Delete a key from service in blocking way.

return Ok(()) even if this key is not exist.

fn scan<'life0, 'life1, 'async_trait>( &'life0 self, path: &'life1 str ) -> Pin<Box<dyn Future<Output = Result<Vec<String>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Scan a key prefix to get all keys that start with this key.

fn blocking_scan(&self, path: &str) -> Result<Vec<String>>

Scan a key prefix to get all keys that start with this key in blocking way.

fn append<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, path: &'life1 str, value: &'life2 [u8] ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Append a key into service

fn blocking_append(&self, path: &str, value: &[u8]) -> Result<()>

Append a key into service in blocking way.