Module opendal::raw::oio

oio provides OpenDAL’s raw traits and types that opendal returns as output.

Those types should only be used internally and we don’t want users to depend on them. So we should also implement trait like AsyncRead for our output traits.

Structs

AppendWriter

AppendWriter will implements [Write] based on append object.

BlockWriter

BlockWriter will implements [Write] based on block uploads.

Entry

Entry is returned by Page or BlockingPage during list operations.

ExactBufWriter

ExactBufWriter is used to implement oio::Write based on exact buffer strategy: flush the underlying storage when the buffered size is exactly the same as the buffer size.

FlatLister

FlatLister will walk dir in bottom up way:

FlexBuf

FlexBuf is a buffer that support frozen bytes and reuse existing allocated memory.

HierarchyLister

ToHierarchyLister will convert a flat list to hierarchy by filter not needed entries.

MultipartPart

The result of MultipartWrite::write_part.

MultipartWriter

MultipartWriter will implements [Write] based on multipart uploads.

OneShotWriter

OneShotWrite is used to implement [Write] based on one shot.

PageContext

PageContext is the context passing between PageList.

PageLister

PageLister implements [List] based on PageList.

PrefixLister

PrefixLister is used to filter entries by prefix.

QueueBuf

QueueBuf is a queue of Buffer.

RangeWriter

RangeWriter will implements [Write] based on range write.

Enums

ListOperation

PageOperation is the name for APIs of lister.

ReadOperation

PageOperation is the name for APIs of lister.

WriteOperation

WriteOperation is the name for APIs of Writer.

Traits

AppendWrite

AppendWrite is used to implement [Write] based on append object. By implementing AppendWrite, services don’t need to care about the details of buffering and uploading parts.

BlockWrite

BlockWrite is used to implement [Write] based on block uploads. By implementing BlockWrite, services don’t need to care about the details of uploading blocks.

BlockingList

BlockingList is the blocking version of List.

BlockingRead

Read is the trait that OpenDAL returns to callers.

BlockingWrite

BlockingWrite is the trait that OpenDAL returns to callers.

List

Page trait is used by raw::Accessor to implement list operation.

MultipartWrite

MultipartWrite is used to implement [Write] based on multipart uploads. By implementing MultipartWrite, services don’t need to care about the details of uploading parts.

OneShotWrite

OneShotWrite is used to implement [Write] based on one shot operation. By implementing OneShotWrite, services don’t need to care about the details.

PageList

PageList is used to implement [List] based on API supporting pagination. By implementing PageList, services don’t need to care about the details of page list.

RangeWrite

RangeWrite is used to implement [Write] based on range write.

Read

Read is the internal trait used by OpenDAL to read data from storage.

Write

Write is the trait that OpenDAL returns to callers.

Type Aliases

BlockingLister

BlockingLister is a boxed BlockingList

BlockingReader

BlockingReader is a boxed dyn BlockingRead.

BlockingWriter

BlockingWriter is a type erased BlockingWrite

Lister

The boxed version of List

Reader

Reader is a type erased Read.

Writer

Writer is a type erased Write