Struct opendal::raw::oio::ChunkedBytes

pub struct ChunkedBytes { /* private fields */ }

ChunkedBytes is used represents a non-contiguous bytes in memory.

Implementations

impl ChunkedBytes

pub fn new() -> Self

Create a new chunked bytes.

pub fn with_chunk_size(chunk_size: usize) -> Self

Create a new chunked cursor with given chunk size.

pub fn from_vec(bs: Vec<Bytes>) -> Self

Build a chunked bytes from a vector of bytes.

This function is guaranteed to run in O(1) time and to not re-allocate the Vec’s buffer or allocate any additional memory.

Reference: https://doc.rust-lang.org/stable/std/collections/struct.VecDeque.html#impl-From%3CVec%3CT,+A%3E%3E-for-VecDeque%3CT,+A%3E

pub fn is_empty(&self) -> bool

Returns true if current cursor is empty.

pub fn len(&self) -> usize

Return current bytes size of cursor.

pub fn clear(&mut self)

Clear the entire cursor.

pub fn push(&mut self, bs: Bytes)

Push a new bytes into ChunkedBytes.

pub fn extend_from_slice(&mut self, bs: &[u8])

Push a new &u8 into ChunkedBytes.

pub fn extend_from_write_buf( &mut self, size: usize, buf: &dyn WriteBuf ) -> usize

Pull data from oio::WriteBuf into ChunkedBytes.