Struct opendal::raw::oio::ExactBufWriter

pub struct ExactBufWriter<W: Write> { /* private fields */ }

ExactBufWriter is used to implement oio::Write based on exact buffer strategy: flush the underlying storage when the buffered size is exactly the same as the buffer size.

ExactBufWriter makes sure that the size of the data written to the underlying storage is exactly buffer_size bytes. It’s useful when the underlying storage requires the size to be written.

For example, R2 requires all parts must be the same size except the last part.

Notes

ExactBufWriter is not a good choice for most cases, because it will cause more network requests.

Implementations

impl<W: Write> ExactBufWriter<W>

pub fn new(inner: W, buffer_size: usize) -> Self

Create a new exact buf writer.