opendal::raw::oio

Trait BlockingWrite

pub trait BlockingWrite:
    Send
    + Sync
    + 'static {
    // Required methods
    fn write(&mut self, bs: Buffer) -> Result<()>;
    fn close(&mut self) -> Result<Metadata>;
}

BlockingWrite is the trait that OpenDAL returns to callers.

Required Methods

fn write(&mut self, bs: Buffer) -> Result<()>

Write whole content at once.

Behavior

Ok(n) means n bytes has been written successfully.

Err(err) means error happens and no bytes has been written.

It’s possible that n < bs.len(), caller should pass the remaining bytes repeatedly until all bytes has been written.

fn close(&mut self) -> Result<Metadata>

Close the writer and make sure all data has been flushed.