Trait opendal::raw::oio::MultipartUploadWrite

pub trait MultipartUploadWrite: Send + Sync + Unpin + 'static {
    // Required methods
    fn write_once<'life0, 'async_trait>(
        &'life0 self,
        size: u64,
        body: AsyncBody
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn initiate_part<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<String>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn write_part<'life0, 'life1, 'async_trait>(
        &'life0 self,
        upload_id: &'life1 str,
        part_number: usize,
        size: u64,
        body: AsyncBody
    ) -> Pin<Box<dyn Future<Output = Result<MultipartUploadPart>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn complete_part<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        upload_id: &'life1 str,
        parts: &'life2 [MultipartUploadPart]
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn abort_part<'life0, 'life1, 'async_trait>(
        &'life0 self,
        upload_id: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

MultipartUploadWrite is used to implement [Write] based on multipart uploads. By implementing MultipartUploadWrite, services don’t need to care about the details of uploading parts.

Architecture

The architecture after adopting MultipartUploadWrite:

Services impl MultipartUploadWrite

MultipartUploadWriter impl Write

Expose MultipartUploadWriter as Accessor::Writer

Notes

MultipartUploadWrite has an oneshot optimization when write has been called only once:

w.write(bs).await?;
w.close().await?;

We will use write_once instead of starting a new multipart upload.

Required Methods

fn write_once<'life0, 'async_trait>( &'life0 self, size: u64, body: AsyncBody ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

write_once is used to write the data to underlying storage at once.

MultipartUploadWriter will call this API when:

All the data has been written to the buffer and we can perform the upload at once.

fn initiate_part<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<String>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

initiate_part will call start a multipart upload and return the upload id.

MultipartUploadWriter will call this when:

the total size of data is unknown.

the total size of data is known, but the size of current write is less then the total size.

fn write_part<'life0, 'life1, 'async_trait>( &'life0 self, upload_id: &'life1 str, part_number: usize, size: u64, body: AsyncBody ) -> Pin<Box<dyn Future<Output = Result<MultipartUploadPart>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

write_part will write a part of the data and returns the result MultipartUploadPart.

MultipartUploadWriter will call this API and stores the result in order.

part_number is the index of the part, starting from 0.

fn complete_part<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, upload_id: &'life1 str, parts: &'life2 [MultipartUploadPart] ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

complete_part will complete the multipart upload to build the final file.

fn abort_part<'life0, 'life1, 'async_trait>( &'life0 self, upload_id: &'life1 str ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

abort_part will cancel the multipart upload and purge all data.