Trait opendal::raw::oio::RangeWrite

pub trait RangeWrite: Send + Sync + Unpin + 'static {
    // Required methods
    fn write_once<'life0, 'async_trait>(
        &'life0 self,
        size: u64,
        body: AsyncBody
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn initiate_range<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<String>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn write_range<'life0, 'life1, 'async_trait>(
        &'life0 self,
        location: &'life1 str,
        written: u64,
        size: u64,
        body: AsyncBody
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn complete_range<'life0, 'life1, 'async_trait>(
        &'life0 self,
        location: &'life1 str,
        written: u64,
        size: u64,
        body: AsyncBody
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn abort_range<'life0, 'life1, 'async_trait>(
        &'life0 self,
        location: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

RangeWrite is used to implement [Write] based on range write.

Services

Services like gcs support range write via GCS Resumable Upload.

GCS will support upload content by specifying the range of the file in CONTENT-RANGE.

Most range based services will have the following limitations:

The size of chunk per upload must be aligned to a certain size. For example, GCS requires to align with 256KiB.

Some services requires to complete the write at the last chunk with the total size.

Architecture

The architecture after adopting RangeWrite:

Services impl RangeWrite

RangeWriter impl Write

Expose RangeWriter as Accessor::Writer

Required Methods

fn write_once<'life0, 'async_trait>( &'life0 self, size: u64, body: AsyncBody ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

write_once is used to write the data to underlying storage at once.

RangeWriter will call this API when:

All the data has been written to the buffer and we can perform the upload at once.

fn initiate_range<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<String>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Initiate range the range write, the returning value is the location.

fn write_range<'life0, 'life1, 'async_trait>( &'life0 self, location: &'life1 str, written: u64, size: u64, body: AsyncBody ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

write_range will write a range of data.

fn complete_range<'life0, 'life1, 'async_trait>( &'life0 self, location: &'life1 str, written: u64, size: u64, body: AsyncBody ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

complete_range will complete the range write by uploading the last chunk.

fn abort_range<'life0, 'life1, 'async_trait>( &'life0 self, location: &'life1 str ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

abort_range will abort the range write by abort all already uploaded data.