Trait opendal::raw::oio::Write

pub trait Write: Unpin + Send + Sync {
    // Required methods
    fn poll_write(
        &mut self,
        cx: &mut Context<'_>,
        bs: &dyn WriteBuf
    ) -> Poll<Result<usize>>;
    fn poll_close(&mut self, cx: &mut Context<'_>) -> Poll<Result<()>>;
    fn poll_abort(&mut self, cx: &mut Context<'_>) -> Poll<Result<()>>;
}

Write is the trait that OpenDAL returns to callers.

Required Methods

fn poll_write( &mut self, cx: &mut Context<'_>, bs: &dyn WriteBuf ) -> Poll<Result<usize>>

Write given bytes into writer.

Behavior

Ok(n) means n bytes has been written successfully.

Err(err) means error happens and no bytes has been written.

It’s possible that n < bs.len(), caller should pass the remaining bytes repeatedly until all bytes has been written.

fn poll_close(&mut self, cx: &mut Context<'_>) -> Poll<Result<()>>

Close the writer and make sure all data has been flushed.

fn poll_abort(&mut self, cx: &mut Context<'_>) -> Poll<Result<()>>

Abort the pending writer.