Trait opendal::raw::oio::WriteBuf

pub trait WriteBuf: Send + Sync {
    // Required methods
    fn remaining(&self) -> usize;
    fn chunk(&self) -> &[u8] ;
    fn advance(&mut self, cnt: usize);

    // Provided methods
    fn copy_to_slice(&self, dst: &mut [u8]) -> usize { ... }
    fn copy_to_bytes(&self, len: usize) -> Bytes { ... }
}

WriteBuf is used in [oio::Write] to provide a trait similar to [bytes::Buf].

The biggest difference is that Buf’s copy_to_slice and copy_to_bytes only needs &self instead of &mut self.

Required Methods

fn remaining(&self) -> usize

Returns the number of bytes between the current position and the end of the buffer.

This value is greater than or equal to the length of the slice returned by chunk().

Notes

Implementations of remaining should ensure that the return value does not change unless a call is made to advance or any other function that is documented to change the Buf’s current position.

fn chunk(&self) -> &[u8]

Returns a slice starting at the current position and of length between 0 and Buf::remaining(). Note that this can return shorter slice (this allows non-continuous internal representation).

Notes

This function should never panic. Once the end of the buffer is reached, i.e., Buf::remaining returns 0, calls to chunk() should return an empty slice.

fn advance(&mut self, cnt: usize)

Advance the internal cursor of the Buf

The next call to chunk() will return a slice starting cnt bytes further into the underlying buffer.

Panics This function may panic if cnt > self.remaining().

Provided Methods

fn copy_to_slice(&self, dst: &mut [u8]) -> usize

Copies current chunk into dst.

Returns the number of bytes copied.

Notes

Users should not assume the returned bytes is the same as the Buf::remaining().

fn copy_to_bytes(&self, len: usize) -> Bytes

Copies current chunk into a bytes.

This function may be optimized by the underlying type to avoid actual copies. For example, Bytes implementation will do a shallow copy (ref-count increment).

Notes

Users should not assume the returned bytes is the same as the Buf::remaining().