Trait opendal::raw::MaybeSend

pub trait MaybeSend: Send { }

MaybeSend is a marker to determine whether a type is Send or not. We use this trait to wrap the Send requirement for wasm32 target.

Safety

MaybeSend is equivalent to Send on non-wasm32 target. And it’s empty trait on wasm32 target to indicate that a type is not Send.