Crate opendal

OpenDAL is the Open Data Access Layer to freely access data.

Documentation: All docs are carried by self, visit docs for more.

Services: All supported services could be found at services.

Layers: All builtin layer could be found at layers.

Features: All features could be found at features.

Quick Start

use opendal::layers::LoggingLayer;
use opendal::services;
use opendal::Operator;
use opendal::Result;

#[tokio::main]
async fn main() -> Result<()> {
    // Pick a builder and configure it.
    let mut builder = services::S3::default();
    builder.bucket("test");

    // Init an operator
    let op = Operator::new(builder)?
        // Init with logging layer enabled.
        .layer(LoggingLayer::default())
        .finish();

    // Write data
    op.write("hello.txt", "Hello, World!").await?;

    // Read data
    let bs = op.read("hello.txt").await?;

    // Fetch metadata
    let meta = op.stat("hello.txt").await?;
    let mode = meta.mode();
    let length = meta.content_length();

    // Delete
    op.delete("hello.txt").await?;

    Ok(())
}

Modules

docs

This module holds documentation for OpenDAL.

layers

Layer is the mechanism to intercept operations.

operator_functions

Functions provides the functions generated by BlockingOperator

operator_futures

Futures provides the futures generated by Operator

raw

Raw modules provide raw APIs that used by underlying services

services

Services will provide builders to build underlying backends.

Structs

Appender

Appender is designed to append data into given path in an asynchronous manner.

BlockingLister

BlockingLister is designed to list entries at given path in a blocking manner.

BlockingOperator

BlockingOperator is the entry for all public blocking APIs.

BlockingReader

BlockingReader is designed to read data from given path in an blocking manner.

BlockingWriter

BlockingWriter is designed to write data into given path in an blocking manner.

Capability

Capability is used to describe what operations are supported by current Operator.

Entry

Entry is the file/dir entry returned by Lister.

Error

Error is the error struct returned by all opendal functions.

Lister

Lister is designed to list entries at given path in an asynchronous manner.

Metadata

Metadata carries all metadata associated with a path.

Operator

Operator is the entry for all public async APIs.

OperatorBuilder

OperatorBuilder is a typed builder to build an Operator.

OperatorInfo

Metadata for operator, users can use this metadata to get information of operator.

Reader

Reader is designed to read data from given path in an asynchronous manner.

Writer

Writer is designed to write data into given path in an asynchronous manner.

Enums

EntryMode

EntryMode represents the mode.

ErrorKind

ErrorKind is all kinds of Error of opendal.

Metakey

Metakey describes the metadata keys that can be stored or queried.

Scheme

Services that OpenDAL supports

Traits

Builder

Builder is used to set up a real underlying service, i.e. storage accessor.

Type Definitions

Result

Result that is a wrapper of Result<T, opendal::Error>