opendal

Module docs

Available on docsrs only.

This module holds documentation for OpenDAL.

It’s highly recommended that you start by reading concepts first.

Modules

changelog

Changes log for all OpenDAL released versions.

comparisons

Compare opendal with other projects to find out the differences and areas that opendal can improve.

concepts

The core concepts of OpenDAL’s public API.

internals

The internal implement details of OpenDAL.

performance

OpenDAL Performance Guide

rfcs

RFCs - OpenDAL Active RFC List

upgrade

Upgrade and migrate procedures while OpenDAL meets breaking changes.