Module opendal::docs::internals::layer

Available on docsrs only.

The internal implementation details of Layer.

Layer itself is quite simple:

pub trait Layer<A: Access> {
    type LayeredAccess: Accessor;

    fn layer(&self, inner: A) -> Self::LayeredAccess;
}

XxxLayer will wrap input Access as inner and return a new Access. So normally the implementation of Layer will be split into two parts:

XxxLayer will implement Layer and return XxxAccessor as Self::LayeredAccess.

XxxAccessor will implement Access and be built by XxxLayer.

Most layer only implements part of Access, so we provide LayeredAccess which will forward all unimplemented methods to inner. It’s highly recommend to implement LayeredAccess trait instead.