Module opendal::docs::rfcs

RFCs power OpenDAL’s development.

To add new features and big refactors:

Start a new RFCs with the template rfc_0000_example.

Submit PR and assign the RFC number with the PR number.

Request reviews from OpenDAL maintainers.

Create a tracking issue and update links in RFC after approval.

Some useful tips:

Start a pre-propose in discussion to communicate quickly.

The proposer of RFC may not be the same person as the implementor. Try to include enough information in RFC itself.

Modules

rfc_0000_example

Proposal Name: (fill me in with a unique ident, my_awesome_feature)

Start Date: (fill me in with today’s date, YYYY-MM-DD)

RFC PR: datafuselabs/opendal#0000

Tracking Issue: datafuselabs/opendal#0000

rfc_0041_object_native_api

Proposal Name: object_native_api

Start Date: 2022-02-18

RFC PR: datafuselabs/opendal#41

Tracking Issue: datafuselabs/opendal#35

rfc_0044_error_handle

Proposal Name: error_handle

Start Date: 2022-02-23

RFC PR: datafuselabs/opendal#44

Tracking Issue: datafuselabs/opendal#43

rfc_0057_auto_region

Proposal Name: auto_region

Start Date: 2022-02-24

RFC PR: datafuselabs/opendal#57

Tracking Issue: datafuselabs/opendal#58

rfc_0069_object_stream

Proposal Name: object_stream

Start Date: 2022-02-25

RFC PR: datafuselabs/opendal#69

Tracking Issue: datafuselabs/opendal#69

rfc_0090_limited_reader

Proposal Name: limited_reader

Start Date: 2022-03-02

RFC PR: datafuselabs/opendal#0090

Tracking Issue: datafuselabs/opendal#0090

rfc_0112_path_normalization

Proposal Name: path-normalization

Start Date: 2022-03-08

RFC PR: datafuselabs/opendal#112

Tracking Issue: datafuselabs/opendal#112

rfc_0191_async_streaming_io

Proposal Name: async_streaming_io

Start Date: 2022-03-28

RFC PR: datafuselabs/opendal#191

Tracking Issue: datafuselabs/opendal#190

rfc_0203_remove_credential

Proposal Name: remove_credential

Start Date: 2022-04-02

RFC PR: datafuselabs/opendal#203

Tracking Issue: datafuselabs/opendal#203

rfc_0221_create_dir

Proposal Name: create-dir

Start Date: 2022-04-06

RFC PR: datafuselabs/opendal#221

Tracking Issue: datafuselabs/opendal#222

rfc_0247_retryable_error

Proposal Name: retryable_error

Start Date: 2022-04-12

RFC PR: datafuselabs/opendal#247

Tracking Issue: datafuselabs/opendal#248

rfc_0293_object_id

Proposal Name: object_id

Start Date: 2022-05-27

RFC PR: datafuselabs/opendal#293

Tracking Issue: datafuselabs/opendal#294

rfc_0337_dir_entry

Proposal Name: dir_entry

Start Date: 2022-06-08

RFC PR: datafuselabs/opendal#337

Tracking Issue: datafuselabs/opendal#338

rfc_0409_accessor_capabilities

Proposal Name: accessor_capabilities

Start Date: 2022-06-29

RFC PR: datafuselabs/opendal#409

Tracking Issue: datafuselabs/opendal#410

rfc_0413_presign

Proposal Name: presign

Start Date: 2022-06-30

RFC PR: datafuselabs/opendal#0413

Tracking Issue: datafuselabs/opendal#394

rfc_0423_command_line_interface

Proposal Name: command_line_interface

Start Date: 2022-07-08

RFC PR: datafuselabs/opendal#423

Tracking Issue: datafuselabs/opendal#422

rfc_0429_init_from_iter

Proposal Name: init_from_iter

Start Date: 2022-07-10

RFC PR: datafuselabs/opendal#429

Tracking Issue: datafuselabs/opendal#430

rfc_0438_multipart

Proposal Name: multipart

Start Date: 2022-07-11

RFC PR: datafuselabs/opendal#438

Tracking Issue: datafuselabs/opendal#439

rfc_0443_gateway

Proposal Name: gateway

Start Date: 2022-07-18

RFC PR: datafuselabs/opendal#443

Tracking Issue: datafuselabs/opendal#444

rfc_0501_new_builder

Proposal Name: new_builder

Start Date: 2022-08-03

RFC PR: datafuselabs/opendal#501

Tracking Issue: datafuselabs/opendal#502

rfc_0554_write_refactor

Proposal Name: write_refactor

Start Date: 2022-08-22

RFC PR: datafuselabs/opendal#554

Tracking Issue: datafuselabs/opendal#555

rfc_0561_list_metadata_reuse

Proposal Name: list_metadata_reuse

Start Date: 2022-08-23

RFC PR: datafuselabs/opendal#561

Tracking Issue: datafuselabs/opendal#570

rfc_0599_blocking_api

Proposal Name: blocking_api

Start Date: 2022-08-30

RFC PR: datafuselabs/opendal#599

Tracking Issue: datafuselabs/opendal#601

rfc_0623_redis_service

Proposal Name: redis_service

Start Date: 2022-08-31

RFC PR: datafuselabs/opendal#0623

Tracking Issue: datafuselabs/opendal#641

rfc_0627_split_capabilities

Proposal Name: split-capabilities

Start Date: 2022-09-04

RFC PR: datafuselabs/opendal#627

Tracking Issue: datafuselabs/opendal#628

rfc_0661_path_in_accessor

Proposal Name: path_in_accessor

Start Date: 2022-09-12

RFC PR: datafuselabs/opendal#661

Tracking Issue: datafuselabs/opendal#662

rfc_0793_generic_kv_services

Proposal Name: generic-kv-services

Start Date: 2022-10-03

RFC PR: datafuselabs/opendal#793

Tracking Issue: datafuselabs/opendal#794

rfc_0926_object_reader

Proposal Name: object_reader

Start Date: 2022-11-13

RFC PR: datafuselabs/opendal#926

Tracking Issue: datafuselabs/opendal#927

rfc_0977_refactor_error

Proposal Name: refactor-error

Start Date: 2022-11-21

RFC PR: datafuselabs/opendal#977

Tracking Issue: datafuselabs/opendal#976

rfc_1085_object_handler

Proposal Name: object_handler

Start Date: 2022-12-19

RFC PR: datafuselabs/opendal#1085

Tracking Issue: datafuselabs/opendal#1085

rfc_1391_object_metadataer

Proposal Name: object_metadataer

Start Date: 2023-02-21

RFC PR: datafuselabs/opendal#1391

Tracking Issue: datafuselabs/opendal#1393

rfc_1398_query_based_metadata

Proposal Name: query_based_metadata

Start Date: 2022-02-22

RFC PR: datafuselabs/opendal#1398

Tracking Issue: datafuselabs/opendal#1398

rfc_1420_object_writer

Proposal Name: object_writer

Start Date: 2023-02-27

RFC PR: datafuselabs/opendal#1420

Tracking Issue: datafuselabs/opendal#1421

rfc_1477_remove_object_concept

Proposal Name: remove_object_concept

Start Date: 2023-03-05

RFC PR: datafuselabs/opendal#1477

Tracking Issue: datafuselabs/opendal#0000