Enum opendal::Scheme

#[non_exhaustive]
pub enum Scheme {

    AliyunDrive,
    Atomicserver,
    Azblob,
    Azdls,
    B2,
    Compfs,
    Seafile,
    Upyun,
    VercelBlob,
    YandexDisk,
    Pcloud,
    Koofr,
    Chainsafe,
    Cacache,
    CloudflareKv,
    Cos,
    D1,
    Dashmap,
    Etcd,
    Foundationdb,
    Dbfs,
    Fs,
    Ftp,
    Gcs,
    Ghac,
    Hdfs,
    Http,
    Huggingface,
    Alluxio,
    Ipfs,
    Ipmfs,
    Icloud,
    Memcached,
    Memory,
    MiniMoka,
    Moka,
    Monoiofs,
    Obs,
    Onedrive,
    Gdrive,
    Dropbox,
    Oss,
    Persy,
    Redis,
    Postgresql,
    Libsql,
    Mysql,
    Sqlite,
    Rocksdb,
    S3,
    Sftp,
    Sled,
    Supabase,
    Swift,
    VercelArtifacts,
    Webdav,
    Webhdfs,
    Redb,
    Tikv,
    Azfile,
    Mongodb,
    Gridfs,
    Github,
    HdfsNative,
    Surrealdb,
    Lakefs,
    NebulaGraph,
    Custom(&'static str),

}

Services that OpenDAL supports

Notes

Scheme is non_exhaustive, new variant COULD be added at any time.

New variant SHOULD be added in alphabet orders,

Users MUST NOT relay on its order.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AliyunDrive

aliyun_drive: Aliyun Drive services.

Atomicserver

atomicserver: Atomicserver services.

Azblob

azblob: Azure Storage Blob services.

Azdls

Azdls: Azure Data Lake Storage Gen2.

B2

B2: Backblaze B2 Services.

Compfs

Compfs: Compio fs Services.

Seafile

Seafile: Seafile Services.

Upyun

Upyun: Upyun Services.

VercelBlob

VercelBlob: VercelBlob Services.

YandexDisk

YandexDisk: YandexDisk Services.

Pcloud

Pcloud: Pcloud Services.

Koofr

Koofr: Koofr Services.

Chainsafe

Chainsafe: Chainsafe Services.

Cacache

cacache: cacache backend support.

CloudflareKv

cloudflare-kv: Cloudflare KV services.

Cos

cos: Tencent Cloud Object Storage services.

D1

d1: D1 services

Dashmap

dashmap: dashmap backend support.

Etcd

etcd: Etcd Services

Foundationdb

foundationdb: Foundationdb services.

Dbfs

dbfs: DBFS backend support.

Fs

fs: POSIX-like file system.

Ftp

ftp: FTP backend.

Gcs

gcs: Google Cloud Storage backend.

Ghac

ghac: GitHub Action Cache services.

Hdfs

hdfs: Hadoop Distributed File System.

Http

http: HTTP backend.

Huggingface

huggingface: Huggingface services.

Alluxio

alluxio: Alluxio services.

Ipfs

ipmfs: IPFS HTTP Gateway

Ipmfs

ipmfs: IPFS mutable file system

Icloud

icloud: APPLE icloud services.

Memcached

memcached: Memcached service support.

Memory

memory: In memory backend support.

MiniMoka

mini-moka: Mini Moka backend support.

Moka

moka: moka backend support.

Monoiofs

monoiofs: monoio fs services.

Obs

obs: Huawei Cloud OBS services.

Onedrive

onedrive: Microsoft OneDrive services.

Gdrive

gdrive: GoogleDrive services.

Dropbox

dropbox: Dropbox services.

Oss

oss: Aliyun Object Storage Services

Persy

persy: persy backend support.

Redis

redis: Redis services

Postgresql

postgresql: Postgresql services

Libsql

libsql: Libsql services

Mysql

mysql: Mysql services

Sqlite

sqlite: Sqlite services

Rocksdb

rocksdb: RocksDB services

S3

s3: AWS S3 alike services.

Sftp

sftp: SFTP services

Sled

sled: Sled services

Supabase

Supabase: Supabase storage service

Swift

swift: Swift backend support.

VercelArtifacts

Vercel Artifacts: Vercel Artifacts service, as known as Vercel Remote Caching.

Webdav

webdav: WebDAV support.

Webhdfs

webhdfs: WebHDFS RESTful API Services

Redb

redb: Redb Services

Tikv

tikv: Tikv Services

Azfile

azfile: Azfile Services

Mongodb

mongodb: MongoDB Services

Gridfs

gridfs: MongoDB Gridfs Services

Github

Github Contents: Github contents support.

HdfsNative

Native HDFS: Hdfs Native service, using rust hdfs-native client for hdfs

Surrealdb

surrealdb: Surrealdb Services

Lakefs

lakefs: LakeFS Services

NebulaGraph

NebulaGraph: NebulaGraph Services

Custom(&'static str)

Custom that allow users to implement services outside of OpenDAL.

NOTE

Custom must not overwrite any existing services name.

Custom must be in lower case.

Implementations

impl Scheme

pub fn into_static(self) -> &'static str

Convert self into static str.

pub fn enabled() -> HashSet<Scheme>

Get all enabled schemes.

OpenDAL could be compiled with different features, which will enable different schemes. This function returns all enabled schemes so users can make decisions based on it.

Examples

use opendal::Scheme;

let enabled_schemes = Scheme::enabled();
if !enabled_schemes.contains(&Scheme::Memory) {
    panic!("s3 support is not enabled")
}