opendal

Enum Scheme

#[non_exhaustive]
pub enum Scheme {

    AliyunDrive,
    Alluxio,
    Azblob,
    Azdls,
    Azfile,
    B2,
    Cacache,
    CloudflareKv,
    Compfs,
    Cos,
    D1,
    Dashmap,
    Dbfs,
    Dropbox,
    Etcd,
    Foundationdb,
    Fs,
    Ftp,
    Gcs,
    Gdrive,
    Ghac,
    Github,
    Gridfs,
    Hdfs,
    HdfsNative,
    Http,
    Huggingface,
    Ipfs,
    Ipmfs,
    Koofr,
    Lakefs,
    Memcached,
    Memory,
    MiniMoka,
    Moka,
    Mongodb,
    Monoiofs,
    Mysql,
    Obs,
    Onedrive,
    Oss,
    Pcloud,
    Persy,
    Postgresql,
    Redb,
    Redis,
    Rocksdb,
    S3,
    Seafile,
    Sftp,
    Sled,
    Sqlite,
    Surrealdb,
    Swift,
    Tikv,
    Upyun,
    VercelArtifacts,
    VercelBlob,
    Webdav,
    Webhdfs,
    YandexDisk,
    Custom(&'static str),

}

Services that OpenDAL supports

Notes

Scheme is non_exhaustive, new variant COULD be added at any time.

New variant SHOULD be added in alphabet orders,

Users MUST NOT relay on its order.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AliyunDrive

aliyun-drive: Aliyun Drive services.

Alluxio

alluxio: Alluxio services.

Azblob

azblob: Azure Storage Blob services.

Azdls

Azdls: Azure Data Lake Storage Gen2.

Azfile

azfile: Azfile Services

B2

B2: Backblaze B2 Services.

Cacache

cacache: cacache backend support.

CloudflareKv

cloudflare-kv: Cloudflare KV services.

Compfs

Compfs: Compio fs Services.

Cos

cos: Tencent Cloud Object Storage services.

D1

d1: D1 services

Dashmap

dashmap: dashmap backend support.

Dbfs

dbfs: DBFS backend support.

Dropbox

dropbox: Dropbox services.

Etcd

etcd: Etcd Services

Foundationdb

foundationdb: Foundationdb services.

Fs

fs: POSIX-like file system.

Ftp

ftp: FTP backend.

Gcs

gcs: Google Cloud Storage backend.

Gdrive

gdrive: GoogleDrive services.

Ghac

ghac: GitHub Action Cache services.

Github

Github Contents: Github contents support.

Gridfs

gridfs: MongoDB Gridfs Services

Hdfs

hdfs: Hadoop Distributed File System.

HdfsNative

Native HDFS: Hdfs Native service, using rust hdfs-native client for hdfs

Http

http: HTTP backend.

Huggingface

huggingface: Huggingface services.

Ipfs

ipmfs: IPFS HTTP Gateway

Ipmfs

ipmfs: IPFS mutable file system

Koofr

Koofr: Koofr Services.

Lakefs

lakefs: LakeFS Services

Memcached

memcached: Memcached service support.

Memory

memory: In memory backend support.

MiniMoka

mini-moka: Mini Moka backend support.

Moka

moka: moka backend support.

Mongodb

mongodb: MongoDB Services

Monoiofs

monoiofs: monoio fs services.

Mysql

mysql: Mysql services

Obs

obs: Huawei Cloud OBS services.

Onedrive

onedrive: Microsoft OneDrive services.

Oss

oss: Aliyun Object Storage Services

Pcloud

Pcloud: Pcloud Services.

Persy

persy: persy backend support.

Postgresql

postgresql: Postgresql services

Redb

redb: Redb Services

Redis

redis: Redis services

Rocksdb

rocksdb: RocksDB services

S3

s3: AWS S3 alike services.

Seafile

Seafile: Seafile Services.

Sftp

sftp: SFTP services

Sled

sled: Sled services

Sqlite

sqlite: Sqlite services

Surrealdb

surrealdb: Surrealdb Services

Swift

swift: Swift backend support.

Tikv

tikv: Tikv Services

Upyun

Upyun: Upyun Services.

VercelArtifacts

Vercel Artifacts: Vercel Artifacts service, as known as Vercel Remote Caching.

VercelBlob

VercelBlob: VercelBlob Services.

Webdav

webdav: WebDAV support.

Webhdfs

webhdfs: WebHDFS RESTful API Services

YandexDisk

YandexDisk: YandexDisk Services.

Custom(&'static str)

Custom that allow users to implement services outside OpenDAL.

NOTE

Custom must not overwrite any existing services name.

Custom must be in lower case.

Implementations

impl Scheme

pub fn into_static(self) -> &'static str

Convert self into static str.

pub fn enabled() -> HashSet<Scheme>

Get all enabled schemes.

OpenDAL could be compiled with different features, which will enable different schemes. This function returns all enabled schemes so users can make decisions based on it.

Examples

use opendal::Scheme;

let enabled_schemes = Scheme::enabled();
if !enabled_schemes.contains(&Scheme::Memory) {
    panic!("s3 support is not enabled")
}