opendal

Module layers

Re-export of layers.

Structs

AsyncBacktraceAccessorlayers-async-backtrace

AsyncBacktraceLayerlayers-async-backtrace

Add Efficient, logical ‘stack’ traces of async functions for the underlying services.

AsyncBacktraceWrapperlayers-async-backtrace

AwaitTreeAccessorlayers-await-tree

AwaitTreeLayerlayers-await-tree

Add an Instrument await-tree for actor-based applications to the underlying services.

AwaitTreeWrapperlayers-await-tree

CapabilityAccessorlayers-capability-check

CapabilityCheckLayerlayers-capability-check

Add an extra capability check layer for every operation

ChaosLayer

Inject chaos into underlying services for robustness test.

ConcurrentLimitLayer

Add concurrent request limit.

DtraceLayer

Support User Statically-Defined Tracing(aka USDT) on Linux

FastmetricsInterceptorlayers-fastmetrics

FastmetricsLayerlayers-fastmetrics

Add fastmetrics for every operation.

FastmetricsLayerBuilderlayers-fastmetrics

FastmetricsLayerBuilder is a config builder to build a FastmetricsLayer.

FastraceAccessorlayers-fastrace

FastraceLayerlayers-fastrace

Add fastrace for every operation.

FastraceWrapperlayers-fastrace

HttpClientLayer

Layer for replacing the default HTTP client with a custom one.

ImmutableDirlayers-immutable-index

ImmutableIndexAccessorlayers-immutable-index

ImmutableIndexLayerlayers-immutable-index

Add an immutable in-memory index for underlying storage services.

LoggingLayer

Add log for every operation.

MetricsInterceptorlayers-metrics

MetricsLayerlayers-metrics

Add metrics for every operation.

MimeGuessAccessorlayers-mime-guess

MimeGuessLayerlayers-mime-guess

A layer that can automatically set Content-Type based on the file extension in the path.

OtelMetricsInterceptorlayers-otel-metrics

OtelMetricsLayerlayers-otel-metrics

Add opentelemetry::metrics for every operation.

OtelMetricsLayerBuilderlayers-otel-metrics

OtelMetricsLayerBuilder is a config builder to build a OtelMetricsLayer.

OtelTraceAccessorlayers-otel-trace

OtelTraceLayerlayers-otel-trace

Add opentelemetry::trace for every operation.

OtelTraceWrapperlayers-otel-trace

PrometheusClientInterceptorlayers-prometheus-client

PrometheusClientLayerlayers-prometheus-client

Add prometheus-client for every operation.

PrometheusClientLayerBuilderlayers-prometheus-client

PrometheusClientLayerBuilder is a config builder to build a PrometheusClientLayer.

PrometheusInterceptorlayers-prometheus

PrometheusLayerlayers-prometheus

Add prometheus for every operation.

PrometheusLayerBuilderlayers-prometheus

PrometheusLayerBuilder is a config builder to build a PrometheusLayer.

RetryLayer

Add retry for temporary failed operations.

SimulateLayer

Simulate missing capabilities for backends in a configurable way.

TailCutAccessorlayers-tail-cut

Accessor that implements tail cut logic.

TailCutLayerlayers-tail-cut

Layer that automatically cancels long-tail requests.

TailCutLayerBuilderlayers-tail-cut

Builder for TailCutLayer.

TailCutWrapperlayers-tail-cut

Wrapper for IO operations (Reader, Writer, Lister, Deleter).

ThrottleAccessorlayers-throttle

ThrottleLayerlayers-throttle

Add a bandwidth rate limiter to the underlying services.

ThrottleWrapperlayers-throttle

TimeoutLayer

Add timeout for every operation to avoid slow or unexpected hang operations.

TracingAccessorlayers-tracing

TracingHttpFetcherlayers-tracing

TracingLayerlayers-tracing

Add tracing for every operation.

TracingStreamlayers-tracing

TracingWrapperlayers-tracing

Traits

LoggingInterceptor

LoggingInterceptor is used to intercept the log.

RetryInterceptor

RetryInterceptor is used to intercept while retry happened.