Module opendal::layers

Layer is the mechanism to intercept operations.

Structs

AsyncBacktraceLayerlayers-async-backtrace

Add Efficient, logical ‘stack’ traces of async functions for the underlying services.

AwaitTreeLayerlayers-await-tree

Add a Instrument await-tree for actor-based applications to the underlying services.

BlockingLayerlayers-blocking

Add blocking API support for non-blocking services.

ChaosLayerlayers-chaos

Inject chaos into underlying services for robustness test.

ConcurrentLimitLayer

Add concurrent request limit.

DtraceLayerLinux and layers-dtrace

Support User Statically-Defined Tracing(aka USDT) on Linux

ImmutableIndexLayer

Add an immutable in-memory index for underlying storage services.

LoggingLayer

Add log for every operations.

MadsimLayerlayers-madsim

Add deterministic simulation for async operations, powered by madsim.

MadsimServerlayers-madsim

A simulated server.This an experimental feature, docs are not ready yet.

MetricsLayerlayers-metrics

Add metrics for every operations.

MinitraceLayerlayers-minitrace

Add minitrace for every operations.

OtelTraceLayerlayers-otel-trace

Add opentelemetry::trace for every operations.

PrometheusClientLayerlayers-prometheus-client

Add prometheus for every operations.

PrometheusLayerlayers-prometheus

Add prometheus for every operations.

RetryLayer

Add retry for temporary failed operations.

ThrottleLayerlayers-throttle

Add a bandwidth rate limiter to the underlying services.

TimeoutLayer

Add timeout for every operations to avoid slow or unexpected hang operations.

TracingLayerlayers-tracing

Add tracing for every operations.

Traits

RetryInterceptor

RetryInterceptor is used to intercept while retry happened.