Struct opendal::operator_futures::OperatorFuture

pub struct OperatorFuture<T, F> { /* private fields */ }

OperatorFuture is the future generated by Operator.

The future will consume all the input to generate a future.

NOTES

This struct is by design to keep in crate. We don’t want users to use this struct directly.

Implementations

impl<F> OperatorFuture<OpStat, F>

pub fn if_match(self, v: &str) -> Self

Set the If-Match for this operation.

pub fn if_none_match(self, v: &str) -> Self

Set the If-None-Match for this operation.

pub fn version(self, v: &str) -> Self

Set the version for this operation.

impl<F> OperatorFuture<(OpStat, Duration), F>

pub fn override_content_disposition(self, v: &str) -> Self

Sets the content-disposition header that should be send back by the remote read operation.

pub fn override_cache_control(self, v: &str) -> Self

Sets the cache-control header that should be send back by the remote read operation.

pub fn override_content_type(self, v: &str) -> Self

Sets the content-type header that should be send back by the remote read operation.

pub fn if_match(self, v: &str) -> Self

Set the If-Match of the option

pub fn if_none_match(self, v: &str) -> Self

Set the If-None-Match of the option

impl<F> OperatorFuture<(OpRead, Duration), F>

pub fn override_content_disposition(self, v: &str) -> Self

Sets the content-disposition header that should be send back by the remote read operation.

pub fn override_cache_control(self, v: &str) -> Self

Sets the cache-control header that should be send back by the remote read operation.

pub fn override_content_type(self, v: &str) -> Self

Sets the content-type header that should be send back by the remote read operation.

pub fn if_match(self, v: &str) -> Self

Set the If-Match of the option

pub fn if_none_match(self, v: &str) -> Self

Set the If-None-Match of the option

impl<F> OperatorFuture<(OpWrite, Duration), F>

pub fn content_type(self, v: &str) -> Self

Set the content type of option

pub fn content_disposition(self, v: &str) -> Self

Set the content disposition of option

pub fn cache_control(self, v: &str) -> Self

Set the content type of option

impl<F> OperatorFuture<(OpRead, BytesRange), F>

pub fn range(self, range: impl RangeBounds<u64>) -> Self

Set the range header for this operation.

pub fn if_match(self, v: &str) -> Self

Set the If-Match for this operation.

pub fn if_none_match(self, v: &str) -> Self

Set the If-None-Match for this operation.

pub fn version(self, v: &str) -> Self

Set the version for this operation.

impl<F> OperatorFuture<OpRead, F>

pub fn version(self, v: &str) -> Self

Set the version for this operation.

impl<F> OperatorFuture<(OpWrite, Bytes), F>

pub fn append(self, v: bool) -> Self

Set the append mode of op.

If the append mode is set, the data will be appended to the end of the file.

Notes

Service could return Unsupported if the underlying storage does not support append.

pub fn buffer(self, v: usize) -> Self

Set the buffer size of op.

If buffer size is set, the data will be buffered by the underlying writer.

NOTE

Service could have their own minimum buffer size while perform write operations like multipart uploads. So the buffer size may be larger than the given buffer size.

pub fn concurrent(self, v: usize) -> Self

Set the maximum concurrent write task amount.

pub fn cache_control(self, v: &str) -> Self

Set the content type of option

pub fn content_type(self, v: &str) -> Self

Set the content type of option

pub fn content_disposition(self, v: &str) -> Self

Set the content disposition of option

impl<F> OperatorFuture<OpWrite, F>

pub fn append(self, v: bool) -> Self

Set the append mode of op.

If the append mode is set, the data will be appended to the end of the file.

Notes

Service could return Unsupported if the underlying storage does not support append.

pub fn buffer(self, v: usize) -> Self

Set the buffer size of op.

If buffer size is set, the data will be buffered by the underlying writer.

NOTE

Service could have their own limitation for buffer size. It’s possible that buffer size is not equal to the given buffer size.

For example:

AWS S3 requires the part size to be in [5MiB, 5GiB].

GCS requires the part size to be aligned with 256 KiB.

The services will alter the buffer size to meet their requirements.

pub fn concurrent(self, v: usize) -> Self

Set the maximum concurrent write task amount.

pub fn cache_control(self, v: &str) -> Self

Set the content type of option

pub fn content_type(self, v: &str) -> Self

Set the content type of option

pub fn content_disposition(self, v: &str) -> Self

Set the content disposition of option

impl<F> OperatorFuture<OpDelete, F>

pub fn version(self, v: &str) -> Self

Change the version of this delete operation.

impl<F> OperatorFuture<OpList, F>

pub fn limit(self, v: usize) -> Self

The limit passed to underlying service to specify the max results that could return per-request.

Users could use this to control the memory usage of list operation.

pub fn start_after(self, v: &str) -> Self

The start_after passes to underlying service to specify the specified key to start listing from.

pub fn recursive(self, v: bool) -> Self

The recursive is used to control whether the list operation is recursive.

If false, list operation will only list the entries under the given path.

If true, list operation will list all entries that starts with given path.

Default to false.

pub fn metakey(self, v: impl Into<FlagSet<Metakey>>) -> Self

Metakey is used to control which meta should be returned.

Lister will make sure the result for specified meta is known:

Some(v) means exist.

None means services doesn’t have this meta.

The default metakey is Metakey::Mode.

pub fn concurrent(self, v: usize) -> Self

Concurrent is used to control the number of concurrent stat requests.

If concurrent is set to <=1, the lister will perform stat requests sequentially.

The default concurrent is 1.