opendal::operator_futures

Type Alias FutureLister

pub type FutureLister<F> = OperatorFuture<OpList, Lister, F>;

Future that generated by Operator::list_with or Operator::lister_with.

Users can add more options by public functions provided by this struct.

Aliased Type

struct FutureLister<F> { /* private fields */ }

Implementations

impl<F: Future<Output = Result<Lister>>> FutureLister<F>

pub fn limit(self, v: usize) -> Self

The limit passed to underlying service to specify the max results that could return per-request.

Users could use this to control the memory usage of list operation.

pub fn start_after(self, v: &str) -> Self

The start_after passes to underlying service to specify the specified key to start listing from.

pub fn recursive(self, v: bool) -> Self

The recursive is used to control whether the list operation is recursive.

If false, list operation will only list the entries under the given path.

If true, list operation will list all entries that starts with given path.

Default to false.

pub fn version(self, v: bool) -> Self

👎Deprecated since 0.51.1: use versions instead

The version is used to control whether the object versions should be returned.

If false, list operation will not return with object versions

If true, list operation will return with object versions if object versioning is supported by the underlying service

Default to false

pub fn versions(self, v: bool) -> Self

Controls whether the list operation should return file versions.

This function allows you to specify if the list operation, when executed, should include information about different versions of files, if versioning is supported and enabled.

If true, subsequent list operations will include version information for each file. If false, version information will be omitted from the list results.

Default to false

pub fn deleted(self, v: bool) -> Self

Controls whether the list operation should include deleted files (or versions).

This function allows you to specify if the list operation, when executed, should include entries for files or versions that have been marked as deleted. This is particularly relevant in object storage systems that support soft deletion or versioning.

If true, subsequent list operations will include deleted files or versions. If false, deleted files or versions will be excluded from the list results.