Module opendal::raw

Raw modules provide raw APIs that used by underlying services

Notes

Only developers who want to develop new services or layers need to access raw APIs.

Raw APIs should only be accessed via opendal::raw::Xxxx, any public API should never expose raw API directly.

Raw APIs are far less stable than public API, please don’t rely on them whenever possible.

Modules

adapters

Providing adapters and its implementations.

oio

oio provides OpenDAL’s raw traits and types that opendal returns as output.

teststests

Utilities for opendal testing.

Structs

AccessorInfo

Metadata for accessor, users can use this metadata to get information of underlying backend.

AtomicContentLength

AtomicContentLength is a wrapper of AtomicU64 that used to store content length.

BytesContentRange

BytesContentRange is the content range of bytes.

BytesRange

BytesRange(offset, size) carries a range of content.

ConcurrentFutures

ConcurrentFutures is a stream that can hold a stream of concurrent futures.

ConcurrentTasks

ConcurrentTasks is used to execute tasks concurrently.

ConfigDeserializer

ConfigDeserializer is used to deserialize given configs from HashMap<String, String>.

FormDataPart

FormDataPart is a builder for multipart/form-data part.

HttpBody

HttpBody is the streaming body that opendal’s HttpClient returned.

HttpClient

HttpClient that used across opendal.

MixedPart

MixedPart is a builder for multipart/mixed part.

Multipart

Multipart is a builder for multipart/form-data.

OpCopy

Args for copy operation.

OpCreateDir

Args for create operation.

OpDelete

Args for delete operation.

OpDeleter

Args for delete operation.

OpList

Args for list operation.

OpPresign

Args for presign operation.

OpRead

Args for read operation.

OpReader

Args for reader operation.

OpRename

Args for rename operation.

OpStat

Args for stat operation.

OpWrite

Args for write operation.

OpWriter

Args for writer operation.

PathCacherinternal-path-cache

PathCacher is a cache for path query.

PresignedRequest

PresignedRequest is a presigned request return by presign.

RpCopy

Reply for copy operation.

RpCreateDir

Reply for create_dir operation

RpDelete

Reply for delete operation

RpList

Reply for list operation.

RpPresign

Reply for presign operation.

RpRead

Reply for read operation.

RpRename

Reply for rename operation.

RpStat

Reply for stat operation.

RpWrite

Reply for write operation.

Enums

FourWays

FourWays is used to implement traits that based on four ways.

Operation

Operation is the name for APIs in Accessor.

PresignOperation

Presign operation used for presign.

ThreeWays

ThreeWays is used to implement traits that based on three ways.

TwoWays

TwoWays is used to implement traits that based on two ways.

Constants

VERSION

VERSION is the compiled version of OpenDAL.

Traits

Access

Underlying trait of all backends for implementers.

AccessDyn

AccessDyn is the dyn version of Access make it possible to use as Box<dyn AccessDyn>.

HttpFetch

HttpFetch is the trait to fetch a request in async way. User should implement this trait to provide their own http client.

Layer

Layer is used to intercept the operations on the underlying storage.

LayeredAccess

LayeredAccess is layered accessor that forward all not implemented method to inner.

MaybeSend

MaybeSend is a marker to determine whether a type is Send or not. We use this trait to wrap the Send requirement for wasm32 target.

Part

Part is a trait for multipart part.

PathQueryinternal-path-cache

The trait required for path cacher.

Functions

build_abs_path

build_abs_path will build an absolute path with root.

build_header_value

Build header value from given string.

build_rel_path

build_rel_path will build a relative path towards root.

build_rooted_abs_path

build_rooted_abs_path will build an absolute path with root.

format_authorization_by_basic

format authorization header by basic auth.

format_authorization_by_bearer

format authorization header by bearer token.

format_content_md5

format content md5 header by given input.

format_datetime_into_http_date

format datetime into http date, this format is required by: https://httpwg.org/specs/rfc9110.html#field.if-modified-since

get_basename

Get basename from path.

get_parent

Get parent from path.

new_json_deserialize_error

Parse json deserialize error into opendal::Error.

new_json_serialize_error

Parse json serialize error into opendal::Error.

new_request_build_error

Create a new error happened during building request.

new_request_credential_error

Create a new error happened during signing request.

new_request_sign_error

Create a new error happened during signing request.

new_std_io_error

Parse std io error into opendal::Error.

new_task_join_errorinternal-tokio-rt

Parse tokio error into opendal::Error.

new_xml_deserialize_error

Parse xml deserialize error into opendal::Error.

normalize_path

Make sure all operation are constructed by normalized path:

normalize_root

Make sure root is normalized to style like /abc/def/.

parse_content_disposition

Parse Content-Disposition for header map

parse_content_encoding

Parse content encoding from header map.

parse_content_length

Parse content length from header map.

parse_content_md5

Parse content md5 from header map.

parse_content_range

Parse content range from header map.

parse_content_type

Parse content type from header map.

parse_datetime_from_from_timestamp

parse datetime from given timestamp

parse_datetime_from_from_timestamp_millis

parse datetime from given timestamp_millis

parse_datetime_from_rfc2822

Parse datetime from rfc2822.

parse_datetime_from_rfc3339

Parse datetime from rfc3339.

parse_etag

Parse etag from header map.

parse_header_to_str

Parse header value to string according to name.

parse_into_metadata

parse_into_metadata will parse standards http headers into Metadata.

parse_last_modified

Parse last modified from header map.

parse_location

Parse redirect location from header map

parse_multipart_boundary

Parse multipart boundary from header map.

parse_prefixed_headers

Parse prefixed headers and return a map with the prefix of each header removed.

percent_decode_path

percent_decode_path will do percent decoding for http decode path.

percent_encode_path

percent_encode_path will do percent encoding for http encode path.

validate_path

Validate given path is match with given EntryMode.

with_error_response_context

Add response context to error.

Type Aliases

Accessor

Accessor is the type erased accessor with Arc<dyn Accessor>.

BoxedFuture

BoxedFuture is the type alias of [futures::future::BoxFuture].

BoxedStaticFuture

BoxedStaticFuture is the type alias of [futures::future::BoxFuture].