opendal::raw

Module adapters

Providing adapters and its implementations.

Adapters in OpenDAL means services that shares similar behaviors. We use adapter to make those services been implemented more easily. For example, with kv::Adapter, users only need to implement get, set for a service.

Notes

Please import the module instead of its type.

For example, use the following:

use opendal::adapters::kv;

impl kv::Adapter for MyType {}

Instead of:

use opendal::adapters::kv::Adapter;

impl Adapter for MyType {}

Available Adapters

kv::Adapter: Adapter for Key Value Services like redis.

typed_kv::Adapter: Adapter key value services that in-memory.

Modules

kv

Providing Key Value Adapter for OpenDAL.

typed_kv

Providing Typed Key Value Adapter for OpenDAL.