Module opendal::raw::adapters::typed_kv

Providing Typed Key Value Adapter for OpenDAL.

Any services that implement Adapter can be used an OpenDAL Service.

Structs

Backend

The typed kv backend which implements Accessor for typed kv adapter.

Capability

Capability is used to describe what operations are supported by Typed KV Operator.

Info

Info for this key value accessor.

Value

Value is the typed value stored in adapter.

Traits

Adapter

Adapter is the typed adapter to underlying kv services.