Module opendal::raw::oio

oio provides OpenDAL’s raw traits and types that opendal returns as output.

Those types should only be used internally and we don’t want users to depend on them. So we should also implement trait like AsyncRead for our output traits.

Structs

AppendObjectWriter

AppendObjectWriter will implements [Write] based on append object.

ByRangeSeekableReader

ByRangeReader that can do seek on non-seekable reader.

ChunkedBytes

ChunkedBytes is used represents a non-contiguous bytes in memory.

CloneableReaderWithinStd

CloneableReaderWithinStd is a Send + Sync + Clone with std::sync::Mutex wrapper of input reader.

CloneableReaderWithinTokio

CloneableReaderWithinTokio is a Send + Sync + Clone with tokio::sync::Mutex wrapper of input reader.

Cursor

Cursor is the cursor for [Bytes] that implements oio::Read

Entry

Entry is returned by Page or BlockingPage during list operations.

ExactBufWriter

ExactBufWriter is used to implement oio::Write based on exact buffer strategy: flush the underlying storage when the buffered size is exactly the same as the buffer size.

FlatPager

ToFlatPager will walk dir in bottom up way:

FromFileReader

FromFileReader is a wrapper of input fd to implement oio::Read.

FromStreamReader

FromStreamReader will convert a futures::Stream<Item = Result<Bytes>> into oio::Read

HierarchyPager

ToHierarchyPager will convert a flat page to hierarchy by filter not needed entries.

MultipartUploadPart

The result of MultipartUploadWrite::write_part.

MultipartUploadWriter

MultipartUploadWriter will implements [Write] based on multipart uploads.

OneShotWriter

OneShotWrite is used to implement [Write] based on one shot.

RangeWriter

RangeWriter will implements [Write] based on range write.

StreamableReader

Make given read streamable.

Enums

PageOperation

PageOperation is the name for APIs of pager.

ReadOperation

PageOperation is the name for APIs of pager.

ThreeWaysWriter

ThreeWaysWriter is used to implement [Write] based on three ways.

TwoWaysWriter

TwoWaysWrite is used to implement [Write] based on two ways.

WriteOperation

WriteOperation is the name for APIs of Writer.

Traits

AppendObjectWrite

AppendObjectWrite is used to implement [Write] based on append object. By implementing AppendObjectWrite, services don’t need to care about the details of buffering and uploading parts.

BlockingPage

BlockingPage is the blocking version of Page.

BlockingRead

Read is the trait that OpenDAL returns to callers.

BlockingWrite

BlockingWrite is the trait that OpenDAL returns to callers.

MultipartUploadWrite

MultipartUploadWrite is used to implement [Write] based on multipart uploads. By implementing MultipartUploadWrite, services don’t need to care about the details of uploading parts.

OneShotWrite

OneShotWrite is used to implement [Write] based on one shot operation. By implementing OneShotWrite, services don’t need to care about the details.

Page

Page trait is used by raw::Accessor to implement list or scan operation.

RangeWrite

RangeWrite is used to implement [Write] based on range write.

Read

Read is the trait that OpenDAL returns to callers.

ReadExt

Extension of Read to make it easier for use.

Stream

Stream is the trait that OpenDAL accepts for sinking data.

StreamExt

Extension of Stream to make it easier for use.

Write

Write is the trait that OpenDAL returns to callers.

WriteBuf

WriteBuf is used in [oio::Write] to provide in-memory buffer support.

WriteExt

Extension of [Read] to make it easier for use.

Functions

into_cloneable_reader_within_std

Convert given reader into a wrapper with std::sync::Mutex for Send + Sync + Clone.

into_cloneable_reader_within_tokio

Convert given reader into a wrapper with tokio::sync::Mutex for Send + Sync + Clone.

into_flat_page

to_flat_pager is used to make a hierarchy pager flat.

into_hierarchy_page

to_hierarchy_pager is used to make a hierarchy pager flat.

into_read_from_file

Convert given file into oio::Reader.

into_read_from_stream

Convert given stream futures::Stream<Item = Result<Bytes>> into oio::Reader.

into_seekable_read_by_range

Convert given reader into oio::Reader by range.

into_stream

Convert given futures stream into oio::Stream.

into_stream_from_reader

Convert given futures reader into oio::Stream.

into_streamable_read

into_streamable is used to make oio::Read or oio::BlockingRead streamable.

Type Definitions

BlockingPager

BlockingPager is a boxed BlockingPage

BlockingReader

BlockingReader is a boxed dyn BlockingRead.

BlockingWriter

BlockingWriter is a type erased BlockingWrite

Pager

The boxed version of Page

Reader

Reader is a type erased Read.

Streamer

Streamer is a type erased Stream.

Writer

Writer is a type erased Write