Module opendal::raw::oio

oio provides OpenDAL’s raw traits and types that opendal returns as output.

Those types should only be used internally and we don’t want users to depend on them. So we should also implement trait like AsyncRead for our output traits.

Structs

AdaptiveBuf

AdaptiveBuf is inspired by hyper ReadStrategy.

AppendWriter

AppendWriter will implements [Write] based on append object.

BlockWriter

BlockWriter will implements [Write] based on block uploads.

BufferReader

BufferReader allows the underlying reader to fetch data at the buffer’s size and is used to amortize the IO’s overhead.

ChunkedBytes

ChunkedBytes is used represents a non-contiguous bytes in memory.

Cursor

Cursor is the cursor for [Bytes] that implements oio::Read

Entry

Entry is returned by Page or BlockingPage during list operations.

ExactBufWriter

ExactBufWriter is used to implement oio::Write based on exact buffer strategy: flush the underlying storage when the buffered size is exactly the same as the buffer size.

FileReader

FileReader that implement range read and streamable read on seekable reader.

FlatLister

FlatLister will walk dir in bottom up way:

FromStreamReader

FromStreamReader will convert a futures::Stream<Item = Result<Bytes>> into oio::Read

FuturesReader

FuturesReader implements oio::Read via [AsyncRead] + [AsyncSeek].

HierarchyLister

ToHierarchyLister will convert a flat list to hierarchy by filter not needed entries.

LazyReader

LazyReader implements oio::Read in a lazy way.

MultipartPart

The result of MultipartWrite::write_part.

MultipartWriter

MultipartWriter will implements [Write] based on multipart uploads.

OneShotWriter

OneShotWrite is used to implement [Write] based on one shot.

PageContext

PageContext is the context passing between PageList.

PageLister

PageLister implements [List] based on PageList.

PrefixLister

PrefixLister is used to filter entries by prefix.

RangeReader

RangeReader that can do seek on non-seekable reader.

RangeWriter

RangeWriter will implements [Write] based on range write.

StdReader

FuturesReader implements oio::BlockingRead via Read + Seek.

StreamableReader

Make given read streamable.

TokioReader

FuturesReader implements oio::Read via [AsyncRead] + [AsyncSeek].

Enums

ListOperation

PageOperation is the name for APIs of lister.

ReadOperation

PageOperation is the name for APIs of lister.

WriteOperation

WriteOperation is the name for APIs of Writer.

Traits

AppendWrite

AppendWrite is used to implement [Write] based on append object. By implementing AppendWrite, services don’t need to care about the details of buffering and uploading parts.

BlockWrite

BlockWrite is used to implement [Write] based on block uploads. By implementing BlockWrite, services don’t need to care about the details of uploading blocks.

BlockingList

BlockingList is the blocking version of List.

BlockingRead

Read is the trait that OpenDAL returns to callers.

BlockingWrite

BlockingWrite is the trait that OpenDAL returns to callers.

List

Page trait is used by raw::Accessor to implement list operation.

ListExt

Extension of List to make it easier for use.

MultipartWrite

MultipartWrite is used to implement [Write] based on multipart uploads. By implementing MultipartWrite, services don’t need to care about the details of uploading parts.

OneShotWrite

OneShotWrite is used to implement [Write] based on one shot operation. By implementing OneShotWrite, services don’t need to care about the details.

PageList

PageList is used to implement [List] based on API supporting pagination. By implementing PageList, services don’t need to care about the details of page list.

RangeWrite

RangeWrite is used to implement [Write] based on range write.

Read

Read is the trait that OpenDAL returns to callers.

ReadExt

Extension of Read to make it easier for use.

Stream

Stream is the trait that OpenDAL accepts for sinking data.

StreamExt

Extension of Stream to make it easier for use.

Write

Write is the trait that OpenDAL returns to callers.

WriteBuf

WriteBuf is used in [oio::Write] to provide in-memory buffer support.

WriteExt

Extension of [Read] to make it easier for use.

Functions

into_read_from_stream

Convert given stream futures::Stream<Item = Result<Bytes>> into oio::Reader.

into_streamNon-WebAssembly

Convert given futures stream into oio::Stream.

into_stream_from_reader

Convert given futures reader into oio::Stream.

into_streamable_read

into_streamable is used to make oio::Read or oio::BlockingRead streamable.

Type Aliases

BlockingLister

BlockingLister is a boxed BlockingList

BlockingReader

BlockingReader is a boxed dyn BlockingRead.

BlockingWriter

BlockingWriter is a type erased BlockingWrite

Lister

The boxed version of List

Reader

Reader is a type erased Read.

Streamer

Streamer is a type erased Stream.

Writer

Writer is a type erased Write