Trait opendal::raw::oio::BlockingRead

pub trait BlockingRead: Send + Sync {
    // Required methods
    fn read(&mut self, buf: &mut [u8]) -> Result<usize>;
    fn seek(&mut self, pos: SeekFrom) -> Result<u64>;
    fn next(&mut self) -> Option<Result<Bytes>>;

    // Provided method
    fn read_to_end(&mut self, buf: &mut Vec<u8>) -> Result<usize> { ... }
}

Read is the trait that OpenDAL returns to callers.

Read is compose of the following trait

Read

Seek

Iterator<Item = Result<Bytes>>

Read is required to be implemented, Seek and Iterator is optional. We use Read to make users life easier.

Required Methods

fn read(&mut self, buf: &mut [u8]) -> Result<usize>

Read synchronously.

fn seek(&mut self, pos: SeekFrom) -> Result<u64>

Seek synchronously.

fn next(&mut self) -> Option<Result<Bytes>>

Iterating [Bytes] from underlying reader.

Provided Methods

fn read_to_end(&mut self, buf: &mut Vec<u8>) -> Result<usize>

Read all data of current reader to the end of buf.