Trait opendal::raw::oio::Read

pub trait Read: Unpin + Send + Sync {
    fn poll_read(
        &mut self,
        cx: &mut Context<'_>,
        buf: &mut [u8]
    ) -> Poll<Result<usize>>;
    fn poll_seek(
        &mut self,
        cx: &mut Context<'_>,
        pos: SeekFrom
    ) -> Poll<Result<u64>>;
    fn poll_next(&mut self, cx: &mut Context<'_>) -> Poll<Option<Result<Bytes>>>;
}

Read is the trait that OpenDAL returns to callers.

Read is compose of the following trait

AsyncRead

AsyncSeek

Stream<Item = Result<Bytes>>

AsyncRead is required to be implemented, AsyncSeek and Stream is optional. We use Read to make users life easier.

Required Methods

fn poll_read(
&mut self,
cx: &mut Context<'_>,
buf: &mut [u8]
) -> Poll<Result<usize>>

Read bytes asynchronously.

fn poll_seek(&mut self, cx: &mut Context<'_>, pos: SeekFrom) -> Poll<Result<u64>>

Seek asynchronously.

Returns Unsupported error if underlying reader doesn’t support seek.

fn poll_next(&mut self, cx: &mut Context<'_>) -> Poll<Option<Result<Bytes>>>

Stream [Bytes] from underlying reader.

Returns Unsupported error if underlying reader doesn’t support stream.

This API exists for avoiding bytes copying inside async runtime. Users can poll bytes from underlying reader and decide when to read/consume them.