Trait opendal::raw::oio::Read

pub trait Read: Unpin + Send + Sync {
    // Required methods
    fn read(
        &mut self,
        limit: usize
    ) -> impl Future<Output = Result<Bytes>> + Send;
    fn seek(
        &mut self,
        pos: SeekFrom
    ) -> impl Future<Output = Result<u64>> + Send;
}

Read is the internal trait used by OpenDAL to read data from storage.

Users should not use or import this trait unless they are implementing an Accessor.

Notes

Object Safety

Read uses async in trait, making it not object safe, preventing the use of Box<dyn Read>. To address this, we’ve introduced [ReadDyn] and its compatible type Box<dyn ReadDyn>.

ReadDyn uses Box::pin() to transform the returned future into a BoxedFuture, introducing an additional layer of indirection and an extra allocation. Ideally, ReadDyn should occur only once, at the outermost level of our API.

Required Methods

fn read(&mut self, limit: usize) -> impl Future<Output = Result<Bytes>> + Send

Available on non-WebAssembly only.

Fetch more bytes from underlying reader.

limit is used to hint the data that user want to read at most. Implementer MUST NOT return more than limit bytes. However, implementer can decide whether to split or merge the read requests underground.

Returning bytes’s length == 0 means:

This reader has reached its “end of file” and will likely no longer be able to produce bytes.

The limit specified was 0.

fn seek(&mut self, pos: SeekFrom) -> impl Future<Output = Result<u64>> + Send

Available on non-WebAssembly only.

Seek asynchronously.

Returns Unsupported error if underlying reader doesn’t support seek.

Object Safety

This trait is not object safe.