Trait opendal::raw::oio::Read

pub trait Read: Unpin + Send + Sync {
    // Required method
    fn read_at(
        &self,
        offset: u64,
        limit: usize
    ) -> impl Future<Output = Result<Buffer>> + Send;
}

Read is the internal trait used by OpenDAL to read data from storage.

Users should not use or import this trait unless they are implementing an Accessor.

Notes

Object Safety

Read uses async in trait, making it not object safe, preventing the use of Box<dyn Read>. To address this, we’ve introduced [ReadDyn] and its compatible type Box<dyn ReadDyn>.

ReadDyn uses Box::pin() to transform the returned future into a BoxedFuture, introducing an additional layer of indirection and an extra allocation. Ideally, ReadDyn should occur only once, at the outermost level of our API.

Required Methods

fn read_at( &self, offset: u64, limit: usize ) -> impl Future<Output = Result<Buffer>> + Send

Available on non-WebAssembly only.

Read at the given offset with the given limit.

Notes

Storage services should try to read as much as possible, only return bytes less than the limit while reaching the end of the file.

Object Safety

This trait is not object safe.