opendal::raw::oio

Trait Write

pub trait Write:
    Unpin
    + Send
    + Sync {
    // Required methods
    fn write(
        &mut self,
        bs: Buffer,
    ) -> impl Future<Output = Result<()>> + MaybeSend;
    fn close(&mut self) -> impl Future<Output = Result<Metadata>> + MaybeSend;
    fn abort(&mut self) -> impl Future<Output = Result<()>> + MaybeSend;
}

Write is the trait that OpenDAL returns to callers.

Required Methods

fn write(&mut self, bs: Buffer) -> impl Future<Output = Result<()>> + MaybeSend

Write given bytes into writer.

Behavior

Ok(()) means all bytes has been written successfully.

Err(err) means error happens and no bytes has been written.

fn close(&mut self) -> impl Future<Output = Result<Metadata>> + MaybeSend

Close the writer and make sure all data has been flushed.

fn abort(&mut self) -> impl Future<Output = Result<()>> + MaybeSend

Abort the pending writer.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.