Trait opendal::raw::oio::Write

pub trait Write: Unpin + Send + Sync {
    // Required methods
    fn write<'life0, 'async_trait>(
        &'life0 mut self,
        bs: Bytes
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn sink<'life0, 'async_trait>(
        &'life0 mut self,
        size: u64,
        s: Streamer
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn abort<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn close<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Write is the trait that OpenDAL returns to callers.

Notes

There are two possible two cases:

Sized: The total size of the object is known in advance.

Unsized: The total size of the object is unknown in advance.

And it’s possible that the given bs length is less than the total content length. Users will call write multiple times to write the whole data.

Required Methods

fn write<'life0, 'async_trait>( &'life0 mut self, bs: Bytes ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Write given bytes into writer.

Notes

It’s possible that the given bs length is less than the total content length. And users will call write multiple times.

Please make sure write is safe to re-enter.

fn sink<'life0, 'async_trait>( &'life0 mut self, size: u64, s: Streamer ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Sink given stream into writer.

fn abort<'life0, 'async_trait>( &'life0 mut self ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Abort the pending writer.

fn close<'life0, 'async_trait>( &'life0 mut self ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Close the writer and make sure all data has been flushed.