Trait opendal::raw::oio::WriteBuf

pub trait WriteBuf: Send + Sync {
    // Required methods
    fn remaining(&self) -> usize;
    fn advance(&mut self, cnt: usize);
    fn chunk(&self) -> &[u8] ;
    fn vectored_chunk(&self) -> Vec<IoSlice<'_>>;
    fn bytes(&self, size: usize) -> Bytes;
    fn vectored_bytes(&self, size: usize) -> Vec<Bytes>;

    // Provided method
    fn is_bytes_optimized(&self, size: usize) -> bool { ... }
}

WriteBuf is used in [oio::Write] to provide in-memory buffer support.

Required Methods

fn remaining(&self) -> usize

Returns the number of bytes between the current position and the end of the buffer.

This value is greater than or equal to the length of the slice returned by chunk().

Notes

Implementations of remaining should ensure that the return value does not change unless a call is made to advance or any other function that is documented to change the Buf’s current position.

fn advance(&mut self, cnt: usize)

Advance the internal cursor of the Buf

The next call to chunk() will return a slice starting cnt bytes further into the underlying buffer.

Panics

This function may panic if cnt > self.remaining().

fn chunk(&self) -> &[u8]

Returns a slice starting at the current position and of length between 0 and Buf::remaining(). Note that this can return shorter slice (this allows non-continuous internal representation).

Notes

This function should never panic. Once the end of the buffer is reached, i.e., Buf::remaining returns 0, calls to chunk() should return an empty slice.

fn vectored_chunk(&self) -> Vec<IoSlice<'_>>

Returns a vectored view of the underlying buffer at the current position and of length between 0 and Buf::remaining(). Note that this can return shorter slice (this allows non-continuous internal representation).

Notes

This function should never panic.

fn bytes(&self, size: usize) -> Bytes

Returns a bytes starting at the current position and of length between 0 and Buf::remaining().

Notes

This functions is used to concat a single bytes from underlying chunks. Use vectored_bytes if you want to avoid copy when possible.

Panics

This function will panic if size > self.remaining().

fn vectored_bytes(&self, size: usize) -> Vec<Bytes>

Returns a vectored bytes of the underlying buffer at the current position and of length between 0 and Buf::remaining().

Notes for Users

This functions is used to return a vectored bytes from underlying chunks. Use bytes if you just want to get a continuous bytes.

Notes for implementors

It’s better to align the vectored bytes with underlying chunks to avoid copy.

Panics

This function will panic if size > self.remaining().

Provided Methods

fn is_bytes_optimized(&self, size: usize) -> bool

Returns true if the underlying buffer is optimized for bytes with given size.

Notes

This function is used to avoid copy when possible. Implementors should return true the given self.bytes(size) could be done without cost. For example, the underlying buffer is Bytes.

Panics

This function will panic if size > self.remaining().