opendal::raw

Module tests

Available on crate feature tests only.

Utilities for opendal testing.

Structs

ReadChecker

ReadChecker is used to check the correctness of the read process.

WriteChecker

WriteAction is used to check the correctness of the write process.

Enums

ReadAction

ReadAction represents a read action.

WriteAction

WriteAction represents a read action.

Statics

TEST_RUNTIME

TEST_RUNTIME is the runtime used for running tests.

Functions

init_test_service

Init a service with given scheme.