opendal

Module services

Services will provide builders to build underlying backends.

More ongoing services support is tracked at opendal#5. Please feel free to submit issues if there are services not covered.

Structs

AliyunDriveservices-aliyun-drive

Capabilities

AliyunDriveConfig

Config for Aliyun Drive services support.

Alluxioservices-alluxio

Alluxio services support.

AlluxioConfig

Config for alluxio services support.

Atomicserverservices-atomicserver

Capabilities

AtomicserverConfig

Config for Atomicserver services support

Azblobservices-azblob

Capabilities

AzblobConfig

Azure Storage Blob services support.

Azdlsservices-azdls

Azure Data Lake Storage Gen2 Support. As known as abfs, azdls or azdls.

AzdlsConfig

Azure Data Lake Storage Gen2 Support.

Azfileservices-azfile

Azure File services support.

AzfileConfig

Azure File services support.

B2services-b2

b2 services support.

B2Config

Config for backblaze b2 services support.

Cacacheservices-cacache

cacache service support.

CacacheConfig

cacache service support.

CloudflareKvservices-cloudflare-kv

Capabilities

CloudflareKvConfig

Cloudflare KV Service Support.

Compfsservices-compfs

[compio]-based file system support.

CompfsConfig

compio-based file system support.

Cosservices-cos

Tencent-Cloud COS services support.

CosConfig

Tencent-Cloud COS services support.

D1services-d1

Capabilities

D1Config

Config for Cloudflare D1 backend support.

Dashmapservices-dashmap

dashmap backend support.

DashmapConfig

dashmap backend support.

Dbfsservices-dbfs

Dbfs’s REST API support. This service will visit the DBFS API supported by Databricks File System.

DbfsConfig

Dbfs’s REST API support.

Dropboxservices-dropbox

Dropbox backend support.

DropboxConfig

Config for Dropbox backend support.

Etcdservices-etcd

Etcd services support.

EtcdConfig

Config for Etcd services support.

Foundationdbservices-foundationdb

Capabilities

FoundationdbConfig

foundationdb service support. Config for FoundationDB.

Fsservices-fs

POSIX file system support.

FsConfig

config for file system

Ftpservices-ftp

FTP and FTPS services support.

FtpConfig

Config for Ftp services support.

Gcsservices-gcs

Google Cloud Storage services support.

GcsConfig

Google Cloud Storage services support.

Gdriveservices-gdrive

GoogleDrive backend support.

GdriveConfig

GoogleDrive configuration.

Ghacservices-ghac

GitHub Action Cache Services support.

GhacConfig

Config for GitHub Action Cache Services support.

Githubservices-github

github contents services support.

GithubConfig

Config for GitHub services support.

Gridfsservices-gridfs

Capabilities

GridfsConfig

Config for Grid file system support.

Hdfsservices-hdfs

A distributed file system that provides high-throughput access to application data.

HdfsConfig

Hadoop Distributed File System (HDFS™) support.

HdfsNativeservices-hdfs-native

A distributed file system that provides high-throughput access to application data. Using Native Rust HDFS client.

HdfsNativeConfig

Config for HdfsNative services support.

Httpservices-http

HTTP Read-only service support like Nginx and Caddy.

HttpConfig

Config for Http service support.

Huggingfaceservices-huggingface

Huggingface’s API support. This service will visit the Huggingface API to access the Huggingface File System. Currently, we only support the model and dataset types of repositories, and operations are limited to reading and listing/stating.

HuggingfaceConfig

Configuration for Huggingface service support.

Icloudservices-icloud

IcloudDrive service support.

IcloudConfig

Config for icloud services support.

Ipfsservices-ipfs

IPFS file system support based on IPFS HTTP Gateway.

IpfsConfig

Config for IPFS file system support.

Ipmfsservices-ipmfs

IPFS file system support based on IPFS MFS API.

IpmfsConfig

Config for IPFS MFS support.

Koofrservices-koofr

Koofr services support.

KoofrConfig

Config for Koofr services support.

Lakefsservices-lakefs

Lakefs’s API support. This service will visit the Lakefs API to access the Lakefs File System. Currently, we only support the model and dataset types of repositories, and operations are limited to reading and listing/stating.

LakefsConfig

Configuration for Lakefs service support.

Memcachedservices-memcached

Memcached service support.

MemcachedConfig

Config for MemCached services support

Memoryservices-memory

In memory service support. (BTreeMap Based)

MemoryConfig

Config for memory.

MiniMokaservices-mini-moka

mini-moka backend support.

MiniMokaConfig

Config for mini-moka support.

Mokaservices-moka

moka backend support.

MokaConfig

Config for Moka services support.

Mongodbservices-mongodb

Capabilities

MongodbConfig

Config for Mongodb service support.

Monoiofsservices-monoiofs

File system support via [monoio].

MonoiofsConfig

Config for monoiofs services support.

Mysqlservices-mysql

Capabilities

MysqlConfig

Config for Mysql services support.

NebulaGraphservices-nebula-graph

Capabilities

NebulaGraphConfig

Config for Mysql services support.

Obsservices-obs

Huawei-Cloud Object Storage Service (OBS) support

ObsConfig

Config for Huawei-Cloud Object Storage Service (OBS) support.

Onedriveservices-onedrive

Microsoft OneDrive backend support.

OnedriveConfig

Config for OneDrive backend support.

Ossservices-oss

Aliyun Object Storage Service (OSS) support

OssConfig

Config for Aliyun Object Storage Service (OSS) support.

Pcloudservices-pcloud

pCloud services support.

PcloudConfig

Config for Pcloud services support.

Persyservices-persy

persy service support.

PersyConfig

Config for persy service support.

Postgresqlservices-postgresql

PostgreSQL services support.

PostgresqlConfig

Config for PostgreSQL services support.

Redbservices-redb

Redb service support.

RedbConfig

Config for redb service support.

Redisservices-redis

Redis services support.

RedisConfig

Config for Redis services support.

Rocksdbservices-rocksdb

RocksDB service support.

RocksdbConfig

Config for Rocksdb Service.

S3services-s3

Aws S3 and compatible services (including minio, digitalocean space, Tencent Cloud Object Storage(COS) and so on) support. For more information about s3-compatible services, refer to Compatible Services.

S3Config

Config for Aws S3 and compatible services (including minio, digitalocean space, Tencent Cloud Object Storage(COS) and so on) support.

Seafileservices-seafile

seafile services support.

SeafileConfig

Config for seafile services support.

Sftpservices-sftp

SFTP services support. (only works on unix)

SftpConfig

Config for Sftp Service support.

Sledservices-sled

Sled services support.

SledConfig

Config for Sled services support.

Sqliteservices-sqlite

Capabilities

SqliteConfig

Config for Sqlite support.

Surrealdbservices-surrealdb

Capabilities

SurrealdbConfig

Config for Surrealdb services support.

Swiftservices-swift

OpenStack Swift’s REST API support. For more information about swift-compatible services, refer to Compatible Services.

SwiftConfig

Config for OpenStack Swift support.

Tikvservices-tikv

TiKV backend builder

TikvConfig

Config for Tikv services support.

Upyunservices-upyun

upyun services support.

UpyunConfig

Config for upyun services support.

VercelArtifactsservices-vercel-artifacts

Vercel Cache backend support.

VercelArtifactsConfig

Config for Vercel Cache support.

VercelBlobservices-vercel-blob

VercelBlob services support.

VercelBlobConfig

Config for VercelBlob services support.

Webdavservices-webdav

WebDAV backend support.

WebdavConfig

Config for WebDAV backend support.

Webhdfsservices-webhdfs

WebHDFS’s REST API support. There two implementations of WebHDFS REST API:

WebhdfsConfig

Config for WebHDFS support.

YandexDiskservices-yandex-disk

YandexDisk services support.

YandexDiskConfig

Config for YandexDisk services support.