opendal

Module services

Services will provide builders to build underlying backends.

More ongoing services support is tracked at opendal#5. Please feel free to submit issues if there are services not covered.

Structs

AliyunDrive

Capabilities

AliyunDriveConfig

Config for Aliyun Drive services support.

Alluxio

Alluxio services support.

AlluxioConfig

Config for alluxio services support.

Azblob

Capabilities

AzblobConfig

Azure Storage Blob services support.

Azdls

Azure Data Lake Storage Gen2 Support. As known as abfs, azdls or azdls.

AzdlsConfig

Azure Data Lake Storage Gen2 Support.

Azfile

Azure File services support.

AzfileConfig

Azure File services support.

B2

b2 services support.

B2Config

Config for backblaze b2 services support.

Cacache

cacache service support.

CacacheConfig

cacache service support.

CloudflareKv

Capabilities

CloudflareKvConfig

Cloudflare KV Service Support.

Compfs

[compio]-based file system support.

CompfsConfig

compio-based file system support.

Cos

Tencent-Cloud COS services support.

CosConfig

Tencent-Cloud COS services support.

D1

Capabilities

D1Config

Config for Cloudflare D1 backend support.

Dashmap

dashmap backend support.

DashmapConfig

Config for Dashmap services support.

Dbfs

Dbfs’s REST API support. This service will visit the DBFS API supported by Databricks File System.

DbfsConfig

Dbfs’s REST API support.

Dropbox

Dropbox backend support.

DropboxConfig

Config for Dropbox backend support.

Etcd

Etcd services support.

EtcdConfig

Config for Etcd services support.

Foundationdb

Capabilities

FoundationdbConfig

foundationdb service support. Config for FoundationDB.

Fs

POSIX file system support.

FsConfig

config for file system

Ftp

FTP and FTPS services support.

FtpConfig

Config for Ftp services support.

Gcs

Google Cloud Storage services support.

GcsConfig

Google Cloud Storage services support.

Gdrive

GoogleDrive backend support.

GdriveConfig

GoogleDrive configuration.

Ghac

GitHub Action Cache Services support.

GhacConfig

Config for GitHub Action Cache Services support.

Github

github contents services support.

GithubConfig

Config for GitHub services support.

Gridfs

Capabilities

GridfsConfig

Config for Grid file system support.

Hdfs

A distributed file system that provides high-throughput access to application data.

HdfsConfig

Hadoop Distributed File System (HDFS™) support.

HdfsNative

Hadoop Distributed File System (HDFS™) support. Using Native Rust HDFS client. A distributed file system that provides high-throughput access to application data. Using Native Rust HDFS client.

HdfsNativeConfig

Config for HdfsNative services support.

Http

HTTP Read-only service support like Nginx and Caddy.

HttpConfig

Config for Http service support.

Huggingface

Huggingface’s API support. This service will visit the Huggingface API to access the Huggingface File System. Currently, we only support the model and dataset types of repositories, and operations are limited to reading and listing/stating.

HuggingfaceConfig

Configuration for Huggingface service support.

Ipfs

IPFS file system support based on IPFS HTTP Gateway.

IpfsConfig

Config for IPFS file system support.

Ipmfs

IPFS file system support based on IPFS MFS API.

IpmfsConfig

Config for IPFS MFS support.

Koofr

Koofr services support.

KoofrConfig

Config for Koofr services support.

Lakefs

Lakefs’s API support. This service will visit the Lakefs API to access the Lakefs File System. Currently, we only support the model and dataset types of repositories, and operations are limited to reading and listing/stating.

LakefsConfig

Configuration for Lakefs service support.

Memcached

Memcached service support.

MemcachedConfig

Config for MemCached services support

Memory

In memory service support. (BTreeMap Based)

MemoryConfig

Config for memory.

MiniMoka

mini-moka backend support.

MiniMokaConfig

Config for mini-moka support.

Moka

moka backend support.

MokaConfig

Config for Moka services support.

MokaValue

Value stored in moka cache containing both metadata and content

Mongodb

Capabilities

MongodbConfig

Config for Mongodb service support.

Monoiofs

File system support via [monoio].

MonoiofsConfig

Config for monoiofs services support.

Mysql

Capabilities

MysqlConfig

Config for Mysql services support.

Obs

Huawei-Cloud Object Storage Service (OBS) support

ObsConfig

Config for Huawei-Cloud Object Storage Service (OBS) support.

Onedrive

Microsoft OneDrive backend support.

OnedriveConfig

Config for OneDrive backend support.

Oss

Aliyun Object Storage Service (OSS) support

OssConfig

Config for Aliyun Object Storage Service (OSS) support.

Pcloud

pCloud services support.

PcloudConfig

Config for Pcloud services support.

Persy

persy service support.

PersyConfig

Config for persy service support.

Postgresql

PostgreSQL services support.

PostgresqlConfig

Config for PostgreSQL services support.

Redb

Redb service support.

RedbConfig

Config for redb service support.

Redis

Redis services support.

RedisConfig

Config for Redis services support.

Rocksdb

RocksDB service support.

RocksdbConfig

Config for Rocksdb Service.

S3

Aws S3 and compatible services (including minio, digitalocean space, Tencent Cloud Object Storage(COS) and so on) support. For more information about s3-compatible services, refer to Compatible Services.

S3Config

Config for Aws S3 and compatible services (including minio, digitalocean space, Tencent Cloud Object Storage(COS) and so on) support.

Seafile

seafile services support.

SeafileConfig

Config for seafile services support.

Sftp

SFTP services support. (only works on unix)

SftpConfig

Config for Sftp Service support.

Sled

Sled services support.

SledConfig

Config for Sled services support.

Sqlite

Capabilities

SqliteConfig

Config for Sqlite support.

Surrealdb

Capabilities

SurrealdbConfig

Config for Surrealdb services support.

Swift

OpenStack Swift’s REST API support. For more information about swift-compatible services, refer to Compatible Services.

SwiftConfig

Config for OpenStack Swift support.

Tikv

TiKV backend builder

TikvConfig

Config for Tikv services support.

Upyun

upyun services support.

UpyunConfig

Config for upyun services support.

VercelArtifacts

Vercel Cache backend support.

VercelArtifactsConfig

Config for Vercel Cache support.

VercelBlob

VercelBlob services support.

VercelBlobConfig

Config for VercelBlob services support.

Webdav

WebDAV backend support.

WebdavConfig

Config for WebDAV backend support.

Webhdfs

WebHDFS’s REST API support. There two implementations of WebHDFS REST API:

WebhdfsConfig

Config for WebHDFS support.

YandexDisk

YandexDisk services support.

YandexDiskConfig

Config for YandexDisk services support.

Constants

ALIYUN_DRIVE_SCHEMEservices-aliyun-drive

Default scheme for aliyun-drive service.

ALLUXIO_SCHEMEservices-alluxio

Default scheme for alluxio service.

AZBLOB_SCHEMEservices-azblob

Default scheme for azblob service.

AZDLS_SCHEMEservices-azdls

Default scheme for azdls service.

AZFILE_SCHEMEservices-azfile

Default scheme for azfile service.

B2_SCHEMEservices-b2

Default scheme for b2 service.

CACACHE_SCHEMEservices-cacache

Default scheme for cacache service.

CLOUDFLARE_KV_SCHEMEservices-cloudflare-kv

Default scheme for cloudflare-kv service.

COMPFS_SCHEMEservices-compfs

Default scheme for compfs service.

COS_SCHEMEservices-cos

Default scheme for cos service.

D1_SCHEMEservices-d1

Default scheme for d1 service.

DASHMAP_SCHEMEservices-dashmap

Default scheme for dashmap service.

DBFS_SCHEMEservices-dbfs

Default scheme for dbfs service.

DROPBOX_SCHEMEservices-dropbox

Default scheme for dropbox service.

ETCD_SCHEMEservices-etcd

Default scheme for etcd service.

FOUNDATIONDB_SCHEMEservices-foundationdb

Default scheme for foundationdb service.

FS_SCHEMEservices-fs

Default scheme for fs service.

FTP_SCHEMEservices-ftp

Default scheme for ftp service.

GCS_SCHEMEservices-gcs

Default scheme for gcs service.

GDRIVE_SCHEMEservices-gdrive

Default scheme for gdrive service.

GHAC_SCHEMEservices-ghac

Default scheme for ghac service.

GITHUB_SCHEMEservices-github

Default scheme for github service.

GRIDFS_SCHEMEservices-gridfs

Default scheme for gridfs service.

HDFS_NATIVE_SCHEMEservices-hdfs-native

Default scheme for hdfs-native service.

HDFS_SCHEMEservices-hdfs

Default scheme for hdfs service.

HTTP_SCHEMEservices-http

Default scheme for http service.

HUGGINGFACE_SCHEMEservices-huggingface

Default scheme for huggingface service.

IPFS_SCHEMEservices-ipfs

Default scheme for ipfs service.

IPMFS_SCHEMEservices-ipmfs

Default scheme for ipmfs service.

KOOFR_SCHEMEservices-koofr

Default scheme for koofr service.

LAKEFS_SCHEMEservices-lakefs

Default scheme for lakefs service.

MEMCACHED_SCHEMEservices-memcached

Default scheme for memcached service.

MEMORY_SCHEMEservices-memory

Default scheme for memory service.

MINI_MOKA_SCHEMEservices-mini-moka

Default scheme for mini-moka service.

MOKA_SCHEMEservices-moka

Default scheme for moka service.

MONGODB_SCHEMEservices-mongodb

Default scheme for mongodb service.

MONOIOFS_SCHEMEservices-monoiofs

Default scheme for monoiofs service.

MYSQL_SCHEMEservices-mysql

Default scheme for mysql service.

OBS_SCHEMEservices-obs

Default scheme for obs service.

ONEDRIVE_SCHEMEservices-onedrive

Default scheme for onedrive service.

OSS_SCHEMEservices-oss

Default scheme for oss service.

PCLOUD_SCHEMEservices-pcloud

Default scheme for pcloud service.

PERSY_SCHEMEservices-persy

Default scheme for persy service.

POSTGRESQL_SCHEMEservices-postgresql

Default scheme for postgresql service.

REDB_SCHEMEservices-redb

Default scheme for redb service.

REDIS_SCHEMEservices-redis

Default scheme for redis service.

ROCKSDB_SCHEMEservices-rocksdb

Default scheme for rocksdb service.

S3_SCHEMEservices-s3

Default scheme for s3 service.

SEAFILE_SCHEMEservices-seafile

Default scheme for seafile service.

SFTP_SCHEMEservices-sftp

Default scheme for sftp service.

SLED_SCHEMEservices-sled

Default scheme for sled service.

SQLITE_SCHEMEservices-sqlite

Default scheme for sqlite service.

SURREALDB_SCHEMEservices-surrealdb

Default scheme for surrealdb service.

SWIFT_SCHEMEservices-swift

Default scheme for swift service.

TIKV_SCHEMEservices-tikv

Default scheme for tikv service.

UPYUN_SCHEMEservices-upyun

Default scheme for upyun service.

VERCEL_ARTIFACTS_SCHEMEservices-vercel-artifacts

Default scheme for vercel-artifacts service.

VERCEL_BLOB_SCHEMEservices-vercel-blob

Default scheme for vercel-blob service.

WEBDAV_SCHEMEservices-webdav

Default scheme for webdav service.

WEBHDFS_SCHEMEservices-webhdfs

Default scheme for webhdfs service.

YANDEX_DISK_SCHEMEservices-yandex-disk

Default scheme for yandex-disk service.

Type Aliases

MokaCacheBuilder

Type alias of moka::future::CacheBuilder