Module opendal::services

Services will provide builders to build underlying backends.

More ongoing services support is tracked at opendal#5. Please feel free to submit issues if there are services not covered.

Structs

Alluxio

Alluxio services support.

AlluxioConfig

Config for alluxio services support.

Atomicserver

Atomicserver service support.

Azblob

Capabilities

AzblobConfig

Azure Storage Blob services support.

Azdls

Azure Data Lake Storage Gen2 Support. As known as abfs, azdls or azdls.

Azfile

Azure File services support.

B2

b2 services support.

B2Config

Config for backblaze b2 services support.

Cacache

cacache service support.

CloudflareKv

Capabilities

Cos

Tencent-Cloud COS services support.

D1

Capabilities

Dashmap

dashmap backend support.

Dbfs

Dbfs’s REST API support. This service will visit the DBFS API supported by Databricks File System.

Dropbox

Dropbox backend support.

Etcd

Etcd services support.

EtcdConfig

Config for Etcd services support.

Foundationdb

foundationdb service support.

Fs

POSIX file system support.

Ftp

FTP and FTPS services support.

FtpConfig

Config for Ftpservices support.

Gcs

Google Cloud Storage services support.

Gdrive

GoogleDrive backend support.

Ghac

GitHub Action Cache Services support.

Gridfs

Capabilities

Hdfs

Hadoop Distributed File System (HDFS™) support. A distributed file system that provides high-throughput access to application data.

Http

HTTP Read-only service support like Nginx and Caddy.

HttpConfig

Config for Http service support.

Huggingface

Huggingface’s API support. This service will visit the Huggingface API to access the Huggingface File System. Currently, we only support the model and dataset types of repositories, and operations are limited to reading and listing/stating.

HuggingfaceConfig

Configuration for Huggingface service support.

Ipfs

IPFS file system support based on IPFS HTTP Gateway.

Ipmfs

IPFS file system support based on IPFS MFS API.

Libsql

Capabilities

LibsqlConfig

Config for Libsqlservices support.

Memcached

Memcached service support.

Memory

In memory service support. (BTreeMap Based)

MiniMoka

mini-moka backend support.

Moka

moka backend support.

MokaConfig

Config for Mokaservices support.

Mongodb

Capabilities

Mysql

Capabilities

MysqlConfig

Config for Mysql services support.

Obs

Huawei-Cloud Object Storage Service (OBS) support

Onedrive

OneDrive backend support.

Oss

Aliyun Object Storage Service (OSS) support

Persy

persy service support.

Postgresql

Postgresql services support.

PostgresqlConfig

Config for PostGresQL services support.

Redb

Redb service support.

Redis

Redis services support.

RedisConfig

Config for Redis services support.

Rocksdb

RocksDB service support.

S3

Aws S3 and compatible services (including minio, digitalocean space, Tencent Cloud Object Storage(COS) and so on) support. For more information about s3-compatible services, refer to Compatible Services.

S3Config

Config for Aws S3 and compatible services (including minio, digitalocean space, Tencent Cloud Object Storage(COS) and so on) support.

Sftp

SFTP services support. (only works on unix)

SftpConfig

Config for Sftpservices support.

Sled

Sled service support.

Sqlite

Capabilities

SqliteConfig

Config for Sqlite support.

Supabase

Supabase service support

Swift

OpenStack Swift’s REST API support. This service will visit the Swift API supported by OpenStack Object Storage.

Tikv

TiKV backend builder

TikvConfig

Config for Tikv services support.

VercelArtifacts

Vercel Cache backend support.

Webdav

WebDAV backend support.

Webhdfs

WebHDFS’s REST API support. There two implementations of WebHDFS REST API: