Struct opendal::Metadata

pub struct Metadata { /* private fields */ }

Metadata carries all metadata associated with a path.

Notes

mode and content_length are required metadata that all services should provide during stat operation. But in list operation, a.k.a., Entry’s content length could be None.

Implementations

impl Metadata

pub fn new(mode: EntryMode) -> Self

Create a new metadata

pub fn mode(&self) -> EntryMode

mode represent this entry’s mode.

pub fn is_file(&self) -> bool

Returns true if this metadata is for a file.

pub fn is_dir(&self) -> bool

Returns true if this metadata is for a directory.

pub fn set_mode(&mut self, v: EntryMode) -> &mut Self

Set mode for entry.

pub fn with_mode(self, v: EntryMode) -> Self

Set mode for entry.

pub fn cache_control(&self) -> Option<&str>

Cache control of this entry. Cache-Control is defined by RFC 7234 Refer to MDN Cache-Control for more information.

pub fn set_cache_control(&mut self, v: &str) -> &mut Self

Set cache control of this entry.

Cache-Control is defined by RFC 7234 Refer to MDN Cache-Control for more information.

pub fn with_cache_control(self, v: String) -> Self

Set cache control of this entry.

Cache-Control is defined by RFC 7234 Refer to MDN Cache-Control for more information.

pub fn content_length(&self) -> u64

Content length of this entry.

Content-Length is defined by RFC 7230 Refer to MDN Content-Length for more information.

pub fn set_content_length(&mut self, v: u64) -> &mut Self

Set content length of this entry.

pub fn with_content_length(self, v: u64) -> Self

Set content length of this entry.

pub fn content_md5(&self) -> Option<&str>

Content MD5 of this entry.

Content MD5 is defined by RFC 2616. And removed by RFC 7231.

OpenDAL will try its best to set this value, but not guarantee this value is the md5 of content.

pub fn set_content_md5(&mut self, v: &str) -> &mut Self

Set content MD5 of this entry.

Content MD5 is defined by RFC 2616. And removed by RFC 7231.

pub fn with_content_md5(self, v: String) -> Self

Set content MD5 of this entry.

Content MD5 is defined by RFC 2616. And removed by RFC 7231.

pub fn content_type(&self) -> Option<&str>

Content Type of this entry.

Content Type is defined by RFC 9110.

pub fn set_content_type(&mut self, v: &str) -> &mut Self

Set Content Type of this entry.

Content Type is defined by RFC 9110.

pub fn with_content_type(self, v: String) -> Self

Set Content Type of this entry.

Content Type is defined by RFC 9110.

pub fn content_range(&self) -> Option<BytesContentRange>

Content Range of this entry.

Content Range is defined by RFC 9110.

pub fn set_content_range(&mut self, v: BytesContentRange) -> &mut Self

Set Content Range of this entry.

Content Range is defined by RFC 9110.

pub fn with_content_range(self, v: BytesContentRange) -> Self

Set Content Range of this entry.

Content Range is defined by RFC 9110.

pub fn last_modified(&self) -> Option<DateTime<Utc>>

Last modified of this entry.

Last-Modified is defined by RFC 7232 Refer to MDN Last-Modified for more information.

OpenDAL parse the raw value into DateTime for convenient.

pub fn set_last_modified(&mut self, v: DateTime<Utc>) -> &mut Self

Set Last modified of this entry.

Last-Modified is defined by RFC 7232 Refer to MDN Last-Modified for more information.

pub fn with_last_modified(self, v: DateTime<Utc>) -> Self

Set Last modified of this entry.

Last-Modified is defined by RFC 7232 Refer to MDN Last-Modified for more information.

pub fn etag(&self) -> Option<&str>

ETag of this entry.

ETag is defined by RFC 7232 Refer to MDN ETag for more information.

OpenDAL will return this value AS-IS like the following:

"33a64df551425fcc55e4d42a148795d9f25f89d4"

W/"0815"

" is part of etag.

pub fn set_etag(&mut self, v: &str) -> &mut Self

Set ETag of this entry.

ETag is defined by RFC 7232 Refer to MDN ETag for more information.

OpenDAL will return this value AS-IS like the following:

"33a64df551425fcc55e4d42a148795d9f25f89d4"

W/"0815"

" is part of etag, don’t trim it before setting.

pub fn with_etag(self, v: String) -> Self

Set ETag of this entry.

ETag is defined by RFC 7232 Refer to MDN ETag for more information.

OpenDAL will return this value AS-IS like the following:

"33a64df551425fcc55e4d42a148795d9f25f89d4"

W/"0815"

" is part of etag, don’t trim it before setting.

pub fn content_disposition(&self) -> Option<&str>

Content-Disposition of this entry

Content-Disposition is defined by RFC 2616 and clarified usage in RFC 6266. Refer to MDN Content-Disposition for more information.

OpenDAL will return this value AS-IS like the following:

“inline”

“attachment”

“attachment; filename="filename.jpg"”

pub fn with_content_disposition(self, v: String) -> Self

Set Content-Disposition of this entry

Content-Disposition is defined by RFC 2616 and clarified usage in RFC 6266. Refer to MDN Content-Disposition for more information.

OpenDAL will return this value AS-IS like the following:

“inline”

“attachment”

“attachment; filename="filename.jpg"”

pub fn set_content_disposition(&mut self, v: &str) -> &mut Self

Set Content-Disposition of this entry

Content-Disposition is defined by RFC 2616 and clarified usage in RFC 6266. Refer to MDN Content-Disposition for more information.

OpenDAL will return this value AS-IS like the following:

“inline”

“attachment”

“attachment; filename="filename.jpg"”