Struct opendal::Writer

pub struct Writer { /* private fields */ }

Writer is designed to write data into given path in an asynchronous manner.

Notes

Please make sure either close or abort has been called before dropping the writer otherwise the data could be lost.

Notes

Writer can be used in two ways:

Sized: write data with a known size by specify the content length.

Unsized: write data with an unknown size, also known as streaming.

All services will support sized writer and provide special optimization if the given data size is the same as the content length, allowing them to be written in one request.

Some services also supports unsized writer. They MAY buffer part of the data and flush them into storage at needs. And finally, the file will be available after close has been called.

Implementations

impl Writer

pub async fn write(&mut self, bs: impl Into<Bytes>) -> Result<()>

Write into inner writer.

pub async fn abort(&mut self) -> Result<()>

Abort the writer and clean up all written data.

Notes

Abort should only be called when the writer is not closed or aborted, otherwise an unexpected error could be returned.

pub async fn close(&mut self) -> Result<()>

Close the writer and make sure all data have been committed.

Notes

Close should only be called when the writer is not closed or aborted, otherwise an unexpected error could be returned.