Trait opendal::Configurator

pub trait Configurator: Serialize + DeserializeOwned + 'static {
    // Required method
    fn into_builder(self) -> impl Builder;

    // Provided method
    fn from_iter(
        iter: impl IntoIterator<Item = (String, String)>,
    ) -> Result<Self> { ... }
}

Configurator is used to configure the underlying service.

This trait allows the developer to define a configuration struct that can:

deserialize from an iterator like hashmap or vector.

convert into a service builder and finally build the underlying services.

Usually, users don’t need to use or import this trait directly, they can use Operator API instead.

For example:

use std::collections::HashMap;

use opendal::services::MemoryConfig;
use opendal::Operator;
async fn test() -> Result<()> {
    let mut cfg = MemoryConfig::default();
    cfg.root = Some("/".to_string());

    // Build an `Operator` to start operating the storage.
    let op: Operator = Operator::from_config(cfg)?.finish();

    Ok(())
}

Required Methods

fn into_builder(self) -> impl Builder

Convert this configuration into a service builder.

Provided Methods

fn from_iter(iter: impl IntoIterator<Item = (String, String)>) -> Result<Self>

Deserialize from an iterator.

This API is provided by opendal, developer should not implement it.

Object Safety

This trait is not object safe.